% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{add.qual.att}
\alias{add.qual.att}
\title{Add a (named) column to a data.frame, filled with the same value}
\usage{
add.qual.att(df, att, val)
}
\arguments{
\item{df}{data.frame}

\item{att}{Name of the new attribute (name of the new column)}

\item{val}{value with which fill the new column}
}
\value{
data.frame
}
\description{
Add a (named) column to a data.frame, filled with the same value
}
