% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{barplotCommonGenesUnique2D}
\alias{barplotCommonGenesUnique2D}
\title{Plot the genes in common between 2 datasets}
\usage{
barplotCommonGenesUnique2D(a, b, name, print = TRUE, publish = TRUE,
  out = "plot", colorpalette, thm, xlab, ylab, baseSize = 12,
  baseFamily = "Linux Libertine")
}
\arguments{
\item{a}{numeric data.frame comprising the first dataset expression data}

\item{b}{numeric data.frame comprising the second dataset expression data}

\item{name}{a length-2 character string vector. Names of the three datasets.}

\item{print}{boolean. Default: TRUE. Whether to print the figure or not.}

\item{publish}{boolean. Default: TRUE. Whether to apply ggplot2::theme_bw to the plot.}

\item{out}{character string. Default: "plot".
Allows to pick the object returned by the function.}

\item{colorpalette}{vector of character strings to personalise the colours of the plot.}

\item{thm}{function of the type of theme(...)}

\item{xlab}{character string. Allows to change the label of the x-axis.}

\item{ylab}{character string. Allows to change the label of the y-axis.}

\item{baseSize}{size of the font. Default: 12. Applied only if "publish"}

\item{baseFamily}{name of the font.
Default is "Linux Libertine". Applied only if "publish"}
}
\value{
depending on the type of "out" can be a data.frame,
        a plot or the data to create the plot
}
\description{
Plot the genes in common between 2 datasets
}
