% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGplots.R
\name{bluered}
\alias{bluered}
\title{Generate blue-to-red colorscale}
\usage{
bluered(n)
}
\arguments{
\item{n}{Desired number of color elements in the panel.}
}
\value{
Vector of HTML-style RGB colors.
}
\description{
Generate blue-to-red colorscale
}
