% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{boolDF2numeric}
\alias{boolDF2numeric}
\title{Convert a logical data.frame to a numeric data.frame (TRUE -> 1; FALSE -> 0)}
\usage{
boolDF2numeric(DF)
}
\arguments{
\item{DF}{a boolean (logical) data.frame}
}
\value{
a numeric data.frame
}
\description{
Convert a logical data.frame to a numeric data.frame (TRUE -> 1; FALSE -> 0)
}
