% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-helpers.R
\name{cat.html}
\alias{cat.html}
\title{Encapsulate result between <pre></pre> html tags}
\usage{
cat.html(string)
}
\arguments{
\item{string}{content to be printed between the <pre></pre> tags}
}
\value{
a bloc of characters
}
\description{
Encapsulate result between <pre></pre> html tags
}
