% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{change.localID}
\alias{change.localID}
\title{Change the name of the samples based on preset annotation tables}
\usage{
change.localID(DF, originalNames, newNames, annot.DF, prefix, suffix)
}
\arguments{
\item{DF}{a data.frame that the columns names have to be changed}

\item{originalNames}{the original column names}

\item{newNames}{the new column names}

\item{annot.DF}{if it exists, takes over originalNames and newNames}

\item{prefix}{string will be add to the colnames after the change of id}

\item{suffix}{string will be removed from the data.frame colnames before the change of id}
}
\value{
a data.frame with changed colnames
}
\description{
Specific to the phd projet
}
