% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{checkConformity}
\alias{checkConformity}
\title{Check that both datasets share identical dimensions}
\usage{
checkConformity(DF1, DF2)
}
\arguments{
\item{DF1}{first data.frame}

\item{DF2}{second data.frame}
}
\value{
summary of the conformity of the two dataset or not
}
\description{
Check that both datasets share identical dimensions
}
