% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{clean.HTSc}
\alias{clean.HTSc}
\title{Remove rows from data.frame (created by older version of irap)
that contain additionnal inforamtion}
\usage{
clean.HTSc(df)
}
\arguments{
\item{df}{data.frame created from older version of irap}
}
\value{
data.frame that contains gene expression only
}
\description{
Remove rows from data.frame (created by older version of irap)
that contain additionnal inforamtion
}
