% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanNames}
\alias{cleanNames}
\title{Take a character string with parenthensis as a pattern and remove it from an another character string}
\usage{
cleanNames(x, pattern)
}
\arguments{
\item{x}{character string that need to be modified}

\item{pattern}{character string to be removed}
}
\value{
a character string
}
\description{
Take a character string with parenthensis as a pattern and remove it from an another character string
}
