% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{cleanReadAnnot}
\alias{cleanReadAnnot}
\title{Allows to load an annotation data.frame}
\usage{
cleanReadAnnot(PATH, standardise = FALSE)
}
\arguments{
\item{PATH}{path of the annotation file}

\item{standardise}{boolean; default: FALSE.
Whether the tissue names should be easier to matched
across datasets.}
}
\value{
a data.frame that contains the annotation data associated to some of the
       dataset used in the PhD project
}
\description{
Allows to load an annotation data.frame
}
