% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-helpers.R
\name{cleantable}
\alias{cleantable}
\title{Wrapper around knitr::kable with HTML as a default}
\usage{
cleantable(..., format = "html")
}
\arguments{
\item{...}{any parameter that can be treated by knitr::kable}

\item{format}{default format="HTML"}
}
\value{
a character vector of the table source code
}
\description{
Wrapper around knitr::kable with HTML as a default
}
\seealso{
\code{\link[knitr]{kable}}
}
