% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{colCbc}
\alias{colCbc}
\title{Wrapper around colorRampalette with two prechosen colours.}
\usage{
colCbc(Colvec = c("aliceblue", "darkcyan"))
}
\arguments{
\item{Colvec}{vector of two string of characters.
Default: c('aliceblue','darkcyan')}
}
\value{
a function
}
\description{
Wrapper around colorRampalette with two prechosen colours.
}
