% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGplots.R
\name{colorpanel}
\alias{colorpanel}
\title{Generate a smoothly varying set of colors}
\usage{
colorpanel(n, low, mid, high)
}
\arguments{
\item{n}{Desired number of color elements in the panel.}

\item{low}{Color to use for the lowest value.}

\item{mid}{Color to use for the  middle value. May be ommited.}

\item{high}{Color to use for the highest values.}
}
\value{
Vector of HTML-style RGB colors.
}
\description{
colorpanel generate a set of colors that varies smoothly. redgreen, greenred, bluered, and redblue generate red-black-green, green-black-red, red-white-blue, and blue-white-red colorbars, respectively. colors
}
