% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-filtering.R
\name{colSelect}
\alias{colSelect}
\title{Selection on a data.frame to keep or filter out specified columns only}
\usage{
colSelect(DF, IN, OUT, ..., type = "DF")
}
\arguments{
\item{DF}{data.frame to be filtered}

\item{IN}{column names to keep}

\item{OUT}{column names to filter out}

\item{...}{other paramaters (for base::grep)}

\item{type}{pick the object to return;
'DF' sends back the filtered dataframe
'names' sends back the names (with the correct order) of the columns that are kept}
}
\value{
a dataframe or an ordered vector of the kept column names
}
\description{
Selection on a data.frame to keep or filter out specified columns only
}
