% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-filtering.R
\name{colSelect2}
\alias{colSelect2}
\title{Selection on a data.frame to keep the specified columns only}
\usage{
colSelect2(DF, IN, ...)
}
\arguments{
\item{DF}{data.frame to filter}

\item{IN}{names of the columns to be kept}

\item{...}{other parameters (for base::grep)}
}
\value{
a filtered dataframe
}
\description{
Selection on a data.frame to keep the specified columns only
}
