% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{comp_ratio_sd}
\alias{comp_ratio_sd}
\title{Calculate the ratio of standard deviation between two data.frames}
\usage{
comp_ratio_sd(DF1, DF2, option1 = TRUE, pseudocount = 1)
}
\arguments{
\item{DF1}{numeric data.frame}

\item{DF2}{numeric data.frame}

\item{option1}{boolean. Default TRUE. Whether to use the first option to compute the ratio sd:
first option: log2(DF1)-log2(DF2); other option: log2(DF1)/log2(DF2)}

\item{pseudocount}{numeric. Default: 1. What to add to avoid problems with log2(0)}
}
\value{
a numeric data.frame
}
\description{
Calculate the ratio of standard deviation between two data.frames
}
