% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptors.R
\name{complement2nbGenesPerCond.unique}
\alias{complement2nbGenesPerCond.unique}
\title{Facetted plot (e.g. per study) of the distribution of genes across the columns.
Genes found uniquely in one condition are highlighted in another colour.}
\usage{
complement2nbGenesPerCond.unique(DF, xlab, ylab, title, legend.title,
  publish = TRUE, threshold = 0, thm, Y, remove.legend = FALSE,
  out = "DF", print = TRUE, facetting = "Study", flip = FALSE)
}
\arguments{
\item{DF}{numerical data.frame. Rownames have to be meaningful. Collates many columns from several studies.}

\item{xlab}{character string; used as x-axis label.}

\item{ylab}{character string; used as y-axis label.}

\item{title}{character string; title of the plot.}

\item{legend.title}{character string; used for the colour legend title}

\item{publish}{boolean. Default: TRUE.}

\item{threshold}{numeric; default: 0. Minimal level of expression to be considered as an expressed gene}

\item{thm}{other theme object (\code{\link[ggplot2]{theme}}) to pass on to the plot}

\item{Y}{numeric vector of two elements. Specify the limites of the y-axis.
'c(minimal limit value, maximal limit value)'}

\item{remove.legend}{boolean. Default: FALSE.}

\item{out}{switch to chose the type of object returned by the function;
default 'DF', i.e. the data.frame with the formatted data used for the plotting.
Other options are:
                   'legend' for the legend of the plot,
                   'plot' for the figure itself,
                   'plot_noLegend' for the plot without the legend,
                   'plot_data' for the ggtable data to recreate the plot
                   'plot_def' for the output of \code{\link[ggplot2]{ggplot_build}} on the plot}

\item{print}{boolean; default: TRUE. Whether the plot should be printed}

\item{facetting}{character string; name of the column that should be used for the facetting.}

\item{flip}{boolean. Default= FALSE. Whether the x-axis and y-axis should be flipped or not.}
}
\value{
depends on out. Objects/plot similar to \code{\link[barzinePhdR]{nbGenesPerCond.unique}}
                       with facetting for the different studies in the original dataframe.
}
\description{
Facetted plot (e.g. per study) of the distribution of genes across the columns.
Genes found uniquely in one condition are highlighted in another colour.
}
