% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{completeDF.t_test.CorrMat_identical.vs.different}
\alias{completeDF.t_test.CorrMat_identical.vs.different}
\title{Apply t.test.CorrMat_identical.vs.different and complete the resulting data.frame}
\usage{
completeDF.t_test.CorrMat_identical.vs.different(CorrMat, label.A, label.B,
  report = TRUE, index = "####", publish = TRUE, QuantMet, datasets,
  CorrelationMethod)
}
\arguments{
\item{CorrMat}{correlation matrix between two studies tissue expression (column correlation)}

\item{label.A}{character string. Name of the first study on which the correlations are based}

\item{label.B}{character string. Name of the second study on which the correlations are based}

\item{report}{boolean. Default: TRUE. Whether to output a text report.}

\item{index}{character string. For the notebook, allows to create new sections.}

\item{publish}{boolean. Default: TRUE. Whether to apply theme_bw()}

\item{QuantMet}{character string. Allows to fill in the quantification of the studies.}

\item{datasets}{Character string. Allows to fill in the names of the studies.}

\item{CorrelationMethod}{character string. Fill in the name of the correlation method}
}
\value{
a data.frame
}
\description{
Apply t.test.CorrMat_identical.vs.different and complete the resulting data.frame
}
