% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{compute_gene_cor}
\alias{compute_gene_cor}
\title{Compute for the genes the correlation of their expression between two datasets
across different tissues}
\usage{
compute_gene_cor(DF1, DF2, method)
}
\arguments{
\item{DF1}{numeric data.frame}

\item{DF2}{numeric data.frame.}

\item{method}{a character string indicating which correlation coefficient is to be computed.
One of "pearson", "spearman", "kendall" or 'pearson_log2'
(for 'pearson_log2' the data is first log2 transformed before the computation of the correlation values)}
}
\value{
a named vector that contains the correlation value of each gene (used as name)
}
\description{
Both data.frames have to share the same number of dimensions.
             For meaningful correlation,
             the rows and the columns should be identical between the two datasets.
}
