% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{compute_gene_cor_save}
\alias{compute_gene_cor_save}
\title{Allows to save to a file the correlation of the genes expression
between two datasets across different tissues}
\usage{
compute_gene_cor_save(DF1, DF2, method, out.df1, out.df2, save, ...)
}
\arguments{
\item{DF1}{numeric data.frame}

\item{DF2}{numeric data.frame}

\item{method}{a character string indicating which correlation coefficient
is to be computed. #' One of "pearson", "kendall", "spearman".}

\item{out.df1}{character string to use for DF1 in the saved file label}

\item{out.df2}{character string to use for DF2 in the saved file label}

\item{save}{boolean; default: TRUE. Whether the correlation should be saved in a file}

\item{...}{other possible parameters for \code{\link[barzinePhdR]{saveToFile}}}
}
\value{
numeric named vector while also saving it to a file
}
\description{
Allows to save to a file the correlation of the genes expression
between two datasets across different tissues
}
