% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{compute_permuted_cor}
\alias{compute_permuted_cor}
\title{Shuffle randomly one expression dataset before computing the correlation with the same labelled rows
from another dataset across the same set of different tissues}
\usage{
compute_permuted_cor(x, y, method = "spearman")
}
\arguments{
\item{x}{numeric data.frame for a first expression study}

\item{y}{numeric data.frame for a second expression study}

\item{method}{a character string indicating which correlation coefficient is to be computed.
One of "pearson", "spearman", "kendall".}
}
\value{
a random permutation for compute_gene_cor
}
\description{
Shuffle randomly one expression dataset before computing the correlation with the same labelled rows
from another dataset across the same set of different tissues
}
