% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{compute_permuted_corQ}
\alias{compute_permuted_corQ}
\title{Optimised to work with clustermq (to be used for LSF jobs)
and shuffle randomly one expression dataset before computing the correlation with the same labelled rows
from another dataset across the same set of different tissues}
\usage{
compute_permuted_corQ(rep, x, y, g.names, method = "spearman",
  random = 145)
}
\arguments{
\item{rep}{positive integer. The current iteration for}

\item{x}{numeric data.frame for a first expression study}

\item{y}{numeric data.frame for a second expression study}

\item{g.names}{character vector. Names of the genes.}

\item{method}{a character string indicating which correlation coefficient is to be computed.
One of "pearson", "spearman", "kendall".}

\item{random}{positive integer. Random seed to add to "rep" so while the shuffling is random,
it can be reproduced.}
}
\value{
a random shuffling instance for compute_gene_cor
}
\description{
Optimised to work with clustermq (to be used for LSF jobs)
and shuffle randomly one expression dataset before computing the correlation with the same labelled rows
from another dataset across the same set of different tissues
}
