% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{compute_tissue_corr_matrix}
\alias{compute_tissue_corr_matrix}
\title{Compute the correlation between two matrices variables( alphabetical order)}
\usage{
compute_tissue_corr_matrix(A, B, label.A, label.B, method, log2 = TRUE,
  report = TRUE, Latex = FALSE, Kable = TRUE, Grid = TRUE,
  headers = TRUE)
}
\arguments{
\item{A}{numeric data.frame for the first study to compare}

\item{B}{numeric data.frame for the second study to compare}

\item{label.A}{character string. Name of the first study used in the report.}

\item{label.B}{character string. Name of the second study used in the report.}

\item{method}{character string. Method for base::cor, either "pearson", "spearman" or
"kendall".}

\item{log2}{logical. Default: TRUE. Whether the data should be transform with log2 or not}

\item{report}{boolean. Default:TRUE. Whether to output the result before returning the matrix.}

\item{Latex}{boolean. Default: FALSE. Whether to output in latex format.}

\item{Kable}{boolean. Default: TRUE. Whether to report with kable.}

\item{Grid}{boolean. Default: TRUE. Whether to create an image of the output.}

\item{headers}{boolean. Default: TRUE. Whether to use headers for the reporting.}
}
\value{
a matrix with the columns correlation of the two studies.
}
\description{
Compute the correlation between two matrices variables( alphabetical order)
}
