% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptors.R
\name{computeBreadth}
\alias{computeBreadth}
\title{Compute breadth of expression of the genes across a data.frame}
\usage{
computeBreadth(DF, threshold, omit.zero, typeR = "vec")
}
\arguments{
\item{DF}{numeric data.frame}

\item{threshold}{numeric. Minimal expression to be considered.
For 0, the comparison is strict.}

\item{omit.zero}{boolean.
Whether the genes which do not reach the minimal expression threshold in any tissue
should be stripped from the output.}

\item{typeR}{character string. Chose the type of result to return.
'vec' for a vector of the observed expression breadths,
'nameVec' for a named vector of the observed expression breadths
associated to each rownames,
'df' or 'DF' for the original data.frame with an additional column,
'dfBool' for a boolean data.frame where TRUE when the gene expression reaches
the minimal expression.
'unique_tissueList' data.frame with genes that are found in one tissue only
 above the minimal threshold.
'plot' for a histogram of the gene expression breadth in that data.frame}
}
\value{
the expression breadth of the genes in different (see typeR)
}
\description{
Compute breadth of expression of the genes across a data.frame
}
