% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-filtering.R
\name{createDFhXg}
\alias{createDFhXg}
\title{Allow to select one tissue only from a data.frame}
\usage{
createDFhXg(DF, Tissue)
}
\arguments{
\item{DF}{a data.frame where the tissue are annotated as a part of the column names}

\item{Tissue}{character string}
}
\value{
a data.frame
}
\description{
Allow to select one tissue only from a data.frame
}
