% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-related.R
\name{createGoAnnot}
\alias{createGoAnnot}
\title{Retrieve GO information for a list of genes or all genes}
\usage{
createGoAnnot(queryList, biomart = "ENSEMBL_MART_ENSEMBL",
  dataset = "hsapiens_gene_ensembl", attributes = c("ensembl_gene_id",
  "name_1006", "go_id"), filters = ("ensembl_gene_id"), host)
}
\arguments{
\item{queryList}{character vector; optional; id of genes}

\item{biomart}{default: 'ENSEMBL_MART_ENSEMBL'}

\item{dataset}{default: 'hsapiens_gene_ensembl'}

\item{attributes}{default: c("ensembl_gene_id","name_1006", "go_id")}

\item{filters}{default; "ensembl_gene_id"}

\item{host}{optional;}
}
\value{
a data.frame with go identifiers and their names.
}
\description{
Retrieve GO information for a list of genes or all genes
}
