% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-helpers.R
\name{createTableResults_basics}
\alias{createTableResults_basics}
\title{Format results output for two-dataframe meta-analysis}
\usage{
createTableResults_basics(DF1, DF2, GeneNameDic, externalSupport,
  select = "all_common", MEAN = TRUE, SD = TRUE, CV = TRUE,
  nb.Tissues = TRUE, MEDIAN = TRUE, MAD = TRUE)
}
\arguments{
\item{DF1}{first data.frame to include}

\item{DF2}{second data.frame to include}

\item{GeneNameDic}{named character vector that maps the gene names to their identifiers}

\item{externalSupport}{allows to include other information about the genes}

\item{select}{'all_common' to keep only the genes that are found in both data.frames
'all' to include all genes from both data.frames}

\item{MEAN}{logical; default: TRUE. Gives the mean value of the gene expression}

\item{SD}{logical; default: TRUE. Gives the standard deviation of the expression of the genes.}

\item{CV}{logical; default: TRUE. Gives the coefficient variation of the expression of the genes}

\item{nb.Tissues}{logical; default: TRUE. The expression breadth of the genes
(i.e. how many tissues the genes has been detected)}

\item{MEDIAN}{logical; default: TRUE. Median expression of the genes}

\item{MAD}{logical; default: TRUE median absolute deviation of the expression of the genes}
}
\value{
a data.frame summarising different aspect of the expression of the genes
}
\description{
Format results output for two-dataframe meta-analysis
}
