% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{creationDf.gxa}
\alias{creationDf.gxa}
\title{Wrapper around read.table with a few tweaks}
\usage{
creationDf.gxa(file, header = TRUE, sep = "\\t",
  stringsAsFactors = FALSE, ..., select = NA, rnames = NA,
  drops = NA)
}
\arguments{
\item{file}{path to the file}

\item{header}{boolean. Default: TRUE. Whether the file contains headers.}

\item{sep}{character. Default: "\ t"}

\item{stringsAsFactors}{boolean. Default=FALSE.}

\item{...}{other arguments for read.table}

\item{select}{Value to keep (?)}

\item{rnames}{default:NA. vector of rownames to keep (?)}

\item{drops}{default:NA. vector of column names that should be removed from the data.frame}
}
\value{
a data.frame
}
\description{
Wrapper around read.table with a few tweaks
}
