% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{cross.select}
\alias{cross.select}
\title{Keep only the columns or rows or both that are common between the two data.frames}
\usage{
cross.select(df1, df2, margin = "c")
}
\arguments{
\item{df1}{data.frame for the first study}

\item{df2}{data.frame for the second study}

\item{margin}{character. Default: 'c'.
Selection only on the columns "c",
selection only on the rows "r",
selection on both "a".}
}
\value{
a list of two data.frames
}
\description{
Keep only the columns or rows or both that are common between the two data.frames
}
