% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{cross.selectCond}
\alias{cross.selectCond}
\title{Selection of the same conditions (columns) between two studies}
\usage{
cross.selectCond(df1, df2, cond = NA, name1 = NA, name2 = NA)
}
\arguments{
\item{df1}{first data.frame}

\item{df2}{second data.frame}

\item{cond}{character string vector.
A list can be given that allows to make a preselection on the kept conditions.}

\item{name1}{character string. Allows to prefix the columns names of the first data.frame}

\item{name2}{character string. Allows to prefix the columns names of the second data.frame}
}
\value{
one data.frame with the selected shared conditions of the two data.frames
}
\description{
Selection of the same conditions (columns) between two studies
}
