% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{cross.sharedDistrib_firstLast}
\alias{cross.sharedDistrib_firstLast}
\title{Draw the figures of the shared expression for the two datasets.}
\usage{
cross.sharedDistrib_firstLast(..., DF1, DF2, omit.zero = TRUE,
  threshold = 0, strict = FALSE, colour, annotate = TRUE,
  publish = TRUE, indexCol)
}
\arguments{
\item{...}{more arguments to be treated}

\item{DF1}{numeric data.frame comprising the expression data of the first study.}

\item{DF2}{numeric data.frame comprising the expression data of the second study.}

\item{omit.zero}{boolean. Default: TRUE.
Whether the genes which do not reach the minimal expression threshold in any tissue
should be stripped from the output.}

\item{threshold}{numeric. Default: 0. Minimal level of expression to be considered as expressed.}

\item{strict}{boolean. Default:FALSE. Whether if only strict equality between the breadth of expression of each gene
between the two studies should be considered
or if the "lapse" argument should be used to define 'similar' cases as well.}

\item{colour}{vector of character strings to personalise the colours of the plot.}

\item{annotate}{boolean. Default: TRUE. Whether to annotate the figures with the counts in each category for an accurate read.}

\item{publish}{boolean. Default: TRUE. Whether to apply ggplot2::theme_bw to the plot.}

\item{indexCol}{vector of integer. Comprise the index of the columns for which the expression breadth has to be compared.}
}
\value{
a list of two figures.
}
\description{
Draw the figures of the shared expression for the two datasets.
}
