% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-filtering.R
\name{crudeStrip}
\alias{crudeStrip}
\title{Strip rows which don't have any value above the given threshold}
\usage{
crudeStrip(DF, threshold)
}
\arguments{
\item{DF}{numeric data.frame}

\item{threshold}{threshold of detection}
}
\value{
a numeric data.frame which can have a smaller number of rows
        than the input data.frame
}
\description{
Strip rows which don't have any value above the given threshold
}
