% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{Cumulative_spe}
\alias{Cumulative_spe}
\title{Rank the genes based on their decreasing order of correlation and then plot them}
\usage{
Cumulative_spe(DF.raw, Title, bin = 100, Table)
}
\arguments{
\item{DF.raw}{numeric data.frame that contains all the correlation}

\item{Title}{character string. Title to figure on the plot.}

\item{bin}{positive integer. eg if N, 1 point out of N points is plotted.}

\item{Table}{data.frame that compiles many information}
}
\value{
a plot
}
\description{
Rank the genes based on their decreasing order of correlation and then plot them
}
