% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{cumulSpe}
\alias{cumulSpe}
\title{Allows to plot the amount of TS genes among the genes that present
the highest observed correlations for gene expression between two studies
once their are ranked by decreasing order.
An additional plot presenting their correlations can be added to ease the interpretation}
\usage{
cumulSpe(cor1, protBreadth, simul1, reference, cor2, simul2,
  labelVec = setNames(c("Protein/mRNA pairs",
  "mRNA/mRNA pairs ('ideal' reference)", "Randomised Protein/mRNA pairs"),
  c("Cor", "reference", "simul1")), relabel = TRUE, sizeLine = 1.1,
  palette, legendText = 0.92, Title, bin = 100, seed = 1323,
  CorrelationPlot = TRUE, intermediaryPlot = FALSE,
  base_family = "Linux Libertine", base_size = 11,
  centerTitle = TRUE, out = "plot")
}
\arguments{
\item{cor1}{named numeric vector that comprises the correlation between two studies.}

\item{protBreadth}{named postive integer that comprises the breadth of expression of the proteins}

\item{simul1}{named numeric vector that comprises the result of the random permutations on cor1}

\item{reference}{named numeric vector that comprises the correlation
between two (transcriptomic) studies that can be used as a reference}

\item{cor2}{named numeric vector that comprises the correlation between two studies.}

\item{simul2}{named numeric vector that comprises the result of the random permutations on cor2}

\item{labelVec}{named character string that allows to name the different categories to be displayed in the legend.
The names should be 'cor1, 'ref', 'simul1', .... and the content the labels
Default: "setNames(c("Protein/mRNA pairs",  "mRNA/mRNA pairs ('ideal' reference)",
          "Randomised Protein/mRNA pairs"), c('Cor','reference','simul1'))"}

\item{relabel}{boolean. Default: TRUE. Whether the categories should be relaballed with the content of labelVec}

\item{sizeLine}{positive numeric. Size of the lines of the plot}

\item{palette}{named vector of character strings. Allows to customise the colours on the plot.
Names should be identical to the categories or to the content of labelVec.}

\item{legendText}{positive numeric. Size of the legend text. Default: 0.92}

\item{Title}{character string. Title of the plot.}

\item{bin}{positive integer.  For the additional plot that uses geom_point(), the interval of ranks to consider between two plotted genes.}

\item{seed}{numeric. Random seed. If simul1 and simul2 are not aggregation of all the random shuffling and one want to plot a specific occurence,
it allows to plot always the same for a given seed}

\item{CorrelationPlot}{boolean. Default:TRUE. Whether to add the correlation plot to help for the interpretation.}

\item{intermediaryPlot}{boolean. Default:FALSE. Additional plot with the percent of TS protein for a given number of (ranked) genes}

\item{base_family}{character string. Name of the font to use. Default: "Linux Libertine"}

\item{base_size}{positive numeric. Size of the font to use as a base.}

\item{centerTitle}{boolean. Default:TRUE. Whether to centre the title or keep it on the left.}

\item{out}{character string. Allows to chose the type of returned object.
"plot" (default) for the plot,
"DF" for the data.frame with all the initial data,
and "annotatedPlot" for an annotated plot}
}
\value{
a plot or a data.frame
}
\description{
Allows to plot the amount of TS genes among the genes that present
the highest observed correlations for gene expression between two studies
once their are ranked by decreasing order.
An additional plot presenting their correlations can be added to ease the interpretation
}
