% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-filtering.R
\name{cutoff}
\alias{cutoff}
\title{Test cell values in a data.frame for equality or higher value than
a specified cutoff}
\usage{
cutoff(DF, CUT)
}
\arguments{
\item{DF}{numeric data.frame}

\item{CUT}{expression cutoff for the test}
}
\value{
a boolean data.frame where
        TRUE means that the value in the original data.frame is higher than CUT.
}
\description{
Test cell values in a data.frame for equality or higher value than
a specified cutoff
}
