% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{dataset.HTSc}
\alias{dataset.HTSc}
\title{Create a data.frame from a folder of htseq outputs}
\usage{
dataset.HTSc(pathFolder, clean = TRUE)
}
\arguments{
\item{pathFolder}{path to the directory}

\item{clean}{boolean. Default: TRUE. Whether the additional annotations should be removed from the result data.frame}
}
\value{
an expression (numeric) data.frame
}
\description{
Create a data.frame from a folder of htseq outputs
}
