% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{dataset.randomize}
\alias{dataset.randomize}
\title{Randomise the content of each column of a data.frame}
\usage{
dataset.randomize(DF, seed = 213)
}
\arguments{
\item{DF}{a data.frame}

\item{seed}{to assure reproducibility. Default:213}
}
\value{
a randomised data.frame
}
\description{
Randomise the content of each column of a data.frame
}
