% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{dataset_irap}
\alias{dataset_irap}
\title{Create a data.frame for a transcriptomic study from the different files outputted by older versions of irap}
\usage{
dataset_irap(path)
}
\arguments{
\item{path}{path to the directory that contains all the files for a given study}
}
\value{
a data.frame
}
\description{
Create a data.frame for a transcriptomic study from the different files outputted by older versions of irap
}
