% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-helpers.R
\name{datatable.custom}
\alias{datatable.custom}
\title{Create a customised HTML table widget based on the DT library}
\usage{
datatable.custom(data, ...)
}
\arguments{
\item{data}{a data object (either a matrix or a data frame)}

\item{...}{other parameters handled by DT::datatable}
}
\value{
an HTML table widget
}
\description{
Create a customised HTML table widget based on the DT library
}
