% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptors.R
\name{density_plot}
\alias{density_plot}
\title{Wrapper around gpplot2's density plot with preset arguments}
\usage{
density_plot(DF, log2 = TRUE, pseudocount = 0, valueLab = "PSM",
  baseFont = "Linux Libertine", condCol, fontSize = 11,
  publish = TRUE, xmin, xmax, ymin, ymax, removeLegend = FALSE, thm,
  alphaScale = 0.2, ylab)
}
\arguments{
\item{DF}{numeric data.frame containing expresion data}

\item{log2}{boolean; default: TRUE.
Whether a log2 transformation should be applied to the values}

\item{pseudocount}{numeric. default: 0.
Whether a pseudocount should be added before the log2 transformation is applied.}

\item{valueLab}{character string representing the unit in which the gene expression is measured;
added}

\item{baseFont}{string character. Name of the font to use for the plot.
Default: "Linux Libertine"}

\item{condCol}{character string vector. Colour palette for the plot.
The vector's names are for the conditions associated to each colour (content of the vector)}

\item{fontSize}{font size. Default: 11.}

\item{publish}{boolean. Default: TRUE. Whether to apply theme_bw() to the plot.}

\item{xmin}{numeric. Lower limit for the x-axis.}

\item{xmax}{numeric. Greater limitefor the x-axis.}

\item{ymin}{numeric. Lower limit for the y-axis.}

\item{ymax}{numeric. Greater limit for the y-axis.}

\item{removeLegend}{boolean, default: FALSE. Whethere to remove the legend}

\item{thm}{function theme( ) with the list of additional elements to apply}

\item{alphaScale}{new alpha level in [0,1]. Default: 0.2}

\item{ylab}{character string. Label for the y-axis}
}
\value{
a density plot
}
\description{
Wrapper around gpplot2's density plot with preset arguments
}
