% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-filtering.R
\name{df.cutoff2}
\alias{df.cutoff2}
\title{Create a data.frame for which the values under a given threshold are set to 0}
\usage{
df.cutoff2(DF, cut, STRIP = FALSE)
}
\arguments{
\item{DF}{numeric data.frame}

\item{cut}{the threshold}

\item{STRIP}{boolean; remove the rows filled with 0 when TRUE}
}
\value{
a numeric data.frame where all (value > cut|| value == 0)
}
\description{
Create a data.frame for which the values under a given threshold are set to 0
}
