% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-filtering.R
\name{df.fill}
\alias{df.fill}
\title{Create a data.frame with new rows or columns for the inputed data.frame
that are filled with a constant}
\usage{
df.fill(DF, NameList, value = 0, type = "row", check = TRUE,
  verbose = TRUE)
}
\arguments{
\item{DF}{data.frame}

\item{NameList}{the header of the new rows or columns}

\item{value}{what the new cells in the data.frame should be filled with}

\item{type}{'row' to create rows otherwise creates columns}

\item{check}{boolean}

\item{verbose}{print the different steps}
}
\value{
a completed data.frame (with the content of Namelist) and filled with the given value
}
\description{
Create a data.frame with new rows or columns for the inputed data.frame
that are filled with a constant
}
