% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-filtering.R
\name{df.Strip}
\alias{df.Strip}
\title{Filter the content of a numeric data.frame based on a threshold}
\usage{
df.Strip(DF, val = 0, margin = 1, case = "gt")
}
\arguments{
\item{DF}{the numeric data.frame}

\item{val}{the value used as a threshold}

\item{margin}{1: for rows, 2 for columns, c(1:2) apply to both column and row.}

\item{case}{'gt' for strictly greater than;
'ge' for greater than or equal to;
'lt' for strictly lesser than;
'le' for lesser than or equal to;
'eq' for equal to}
}
\value{
Data.frame which exclusively contains rows that comply at least once to the given criterion
}
\description{
Filter the content of a numeric data.frame based on a threshold
}
