% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{drawOverlapHeatmaps}
\alias{drawOverlapHeatmaps}
\title{Draw the heatmap based on the output of matrix overlap_res}
\usage{
drawOverlapHeatmaps(dfPercent, dfpvalue, label.df1, label.df2,
  col = barzinePhdR::colCbc(), colp)
}
\arguments{
\item{dfPercent}{overlap matrix computed by matrix overlap_res}

\item{dfpvalue}{associated p-value matrix}

\item{label.df1}{character string. Label for the first data.frame}

\item{label.df2}{character string. Label for the second data.frame}

\item{col}{colour palette for the main heatmap}

\item{colp}{colour palette for the p-value associated heatmap.}
}
\value{
two heatmaps one with the percentage of overlap between the two studies
        and the second with the corresponding p-values.
}
\description{
Draw the heatmap based on the output of matrix overlap_res
}
