% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiag.R
\name{DrawVenn}
\alias{DrawVenn}
\title{Function based on the VennDiagram package with some customisations.
Data.frames are accepted, the venn diagram compares then the colnames or rownames
Allows the use of non-ascii characters}
\usage{
DrawVenn(..., a, b, c, d, e, names, fills, cols, titlelab, fig = TRUE,
  type = "vec", titlesize, xlwd = 2, xlty = "solid", xalpha = 0.3)
}
\arguments{
\item{...}{Any non-explicit parameter of the function accepted by VennDiagram::venn.diagram}

\item{a}{First set}

\item{b}{Second set}

\item{c}{Third set}

\item{d}{Fourth set}

\item{e}{Fifth set}

\item{names}{Names that will be outputed on the plot}

\item{fills}{Colors associated with each set which fill the areas}

\item{cols}{Colors associated with each set which will delimit the areas}

\item{titlelab}{Title of the plot}

\item{fig}{boolean; whether if the plot should be directly printed or just sent back as an object}

\item{type}{How the data is structured and should be compared.
vec: comparison occurs on the content of the respective vectors;
namedV: the names of the vector should be compared;
rownames: the row.names of the respective data.frames are used for the comparison;
colnames: the col.names of the respective data.frames are used for the comparison.}

\item{titlesize}{Size of the title}

\item{xlwd}{Numeric vector giving the width of each circle's circumference}

\item{xlty}{Numeric vector giving the dash pattern of each circle's circumference}

\item{xalpha}{Numeric vector giving the alpha transparency of each circle's area}
}
\value{
sends back a plot object, which can be directly printed as well.
}
\description{
Function based on the VennDiagram package with some customisations.
Data.frames are accepted, the venn diagram compares then the colnames or rownames
Allows the use of non-ascii characters
}
