% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{ebiRank}
\alias{ebiRank}
\title{Rank expression of the genes within each tissue.}
\usage{
ebiRank(DF)
}
\arguments{
\item{DF}{numeric data.frame}
}
\value{
numeric data.frame
}
\description{
Rank expression of the genes within each tissue.
}
