% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{enhancedTissue}
\alias{enhancedTissue}
\title{Give the list of the genes whose expression is enhanced for a tissue at the given threshold.}
\usage{
enhancedTissue(DF, fold = 5, detect = FALSE, threshold = 1,
  clean = TRUE)
}
\arguments{
\item{DF}{numeric data.frame}

\item{fold}{numeric; default: 5. Above which fold change a gene is considered as enhanced for a tissue}

\item{detect}{numeric; default:0. Minimal level of expression to be considered detected}

\item{threshold}{numeric; default:1. Minimal level of expression to be considered as possibly translated as protein}

\item{clean}{logical; default: TRUE. Remove the genes that haven't show any enhancement}
}
\value{
named vector. Names give the gene ID (from the rownames)
        and the content is the tissue where the enhancement has been observed
}
\description{
Give the list of the genes whose expression is enhanced for a tissue at the given threshold.
}
