% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{enrichedTissue}
\alias{enrichedTissue}
\title{Give the list of the genes whose expression is enriched for a tissue at the given threshold.}
\usage{
enrichedTissue(DF, fold = 5, detect = FALSE, threshold = 1,
  verbose = TRUE, clean = TRUE)
}
\arguments{
\item{DF}{numeric data.frame}

\item{fold}{numeric. default: 5. Above which fold change a gene is considered as enriched for a tissue}

\item{detect}{numeric; default:0. Minimal level of expression to be considered detected}

\item{threshold}{numeric; default:1. Minimal level of expression to be considered as possibly translated as protein}

\item{verbose}{logical; default:TRUE. Print to screen the number of genes that are enriched and aren't}

\item{clean}{logical; default: TRUE. Remove the genes that haven't show any enrichment}
}
\value{
named vector. Names give the gene ID (from the rownames)
        and the content is the tissue where the enrichment has been observed
}
\description{
Give the list of the genes whose expression is enriched for a tissue at the given threshold.
}
