% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{evolCorrCumul}
\alias{evolCorrCumul}
\title{Plot the correlation of the gene expression between the same tissues of two studies
as a function of the number of genes included in the correlation computation which
varies based on the minimal threshold of expression of the genes.}
\usage{
evolCorrCumul(DFa, sepSeq, step = 10, method = "pearson",
  use = "pairwise.complete.obs", filename,
  fontfamily = "Linux Libertine", fontSize = 11, simple = FALSE,
  sep = "(", verbose = FALSE, position.legend = "bottom",
  legendN = 2, point = TRUE, smooth = FALSE, line = FALSE,
  noTitleLegend = TRUE, Colorpalette, logscaleX = FALSE, ylabTitle,
  shorten = FALSE, Title, unit)
}
\arguments{
\item{DFa}{data.frame that contains the expression data for one tissue across several studies}

\item{sepSeq}{positive integer vector. Give the indices of the genes for which the expression
should be plotted once ordered.}

\item{step}{positive integer. The step to use between the rank to plot.}

\item{method}{a character string indicating which correlation coefficient is to be computed.
One of "pearson", "kendall" or "spearman".}

\item{use}{character string which allows to pick the way missing data are handled.
one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs".}

\item{filename}{character string. Path where the figure should be saved.}

\item{fontfamily}{character string, name of the font to use. Default: "Linux Libertine"}

\item{fontSize}{positive numeric. Size of the font to use as a base.}

\item{simple}{boolean. Default: FALSE. Whether to simplify the plot.}

\item{sep}{character string. Default: '('. Allows to retrieve the name of the study from the colnames
(preformated colnames)}

\item{verbose}{boolean. Default: FALSE. output additional messages.}

\item{position.legend}{character string. Where to position the legend in regard of the plot.
Default: "bottom", can be also 'right', 'left' and 'top'.}

\item{legendN}{positive integer. Allows to apply a number of rows to the legend.}

\item{point}{boolean. Default: TRUE. Whether to draw the curves as a separate point for each value}

\item{smooth}{boolean. Default:FALSE. Whether to add a smooth line ("lm" method)}

\item{line}{boolean. Default:FALSE. Whether to draw the curves as lines.}

\item{noTitleLegend}{boolean. Default: TRUE. Whether to remove the title legend.}

\item{Colorpalette}{Named vector of character strings that allows to customise the colours in the plot}

\item{logscaleX}{boolean. Default: FALSE. Whether to apply a log10 transformation to the x-axis scale.}

\item{ylabTitle}{character string. Allows to redefine the label of the y-axis.}

\item{shorten}{boolean. Default: FALSE. Allows to shorten the names of the plotted objects (columns)}

\item{Title}{character string. Title of the plot}

\item{unit}{character string. Allows to fill in the unit in which the genes are quantified}
}
\value{
a figure
}
\description{
Plot the correlation of the gene expression between the same tissues of two studies
as a function of the number of genes included in the correlation computation which
varies based on the minimal threshold of expression of the genes.
}
