% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{evolCvCorrCumul}
\alias{evolCvCorrCumul}
\title{Plot the correlation of the gene expression
while taking in account the coefficient of variation of the genes
between the same tissues of two studies
as a function of the number of genes included in the correlation computation which
varies based on the minimal threshold of expression of the genes.}
\usage{
evolCvCorrCumul(DFa, cvDF, sepSeq, step = 0.01, method = "pearson",
  use = "pairwise.complete.obs", filename,
  fontfamily = "Linux Libertine", Colorpalette, legend = FALSE,
  point = FALSE, line = FALSE, smooth = TRUE)
}
\arguments{
\item{DFa}{data.frame that contains the expression data for one tissue across several studies}

\item{cvDF}{data.frame that contains the coefficient of variation of the different genes across the tissues}

\item{sepSeq}{positive integer vector. Give the indices of the genes for which the expression
should be plotted once ordered.}

\item{step}{positive integer. The step to use between the rank to plot.}

\item{method}{a character string indicating which correlation coefficient is to be computed.
One of "pearson", "kendall" or "spearman".}

\item{use}{character string which allows to pick the way missing data are handled.
one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs".}

\item{filename}{character string. Path where the figure should be saved.}

\item{fontfamily}{character string, name of the font to use. Default: "Linux Libertine"}

\item{Colorpalette}{Named vector of character strings that allows to customise the colours in the plot}

\item{legend}{boolean. Default: FALSE. Whether the legend should be added to the plot.}

\item{point}{boolean. Default: FALSE. Whether to plot with geom_point()}

\item{line}{boolean. Default: FALSE. Whether to plot with geom_line()}

\item{smooth}{boolean. Default: TRUE. Whether to use a smooth line.}
}
\value{
a plot
}
\description{
Plot the correlation of the gene expression
while taking in account the coefficient of variation of the genes
between the same tissues of two studies
as a function of the number of genes included in the correlation computation which
varies based on the minimal threshold of expression of the genes.
}
