% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-genesExpDF.R
\name{excludeGenes}
\alias{excludeGenes}
\title{Exclude the genes (rows) from a data.frame based on a given list}
\usage{
excludeGenes(DF, genelist)
}
\arguments{
\item{DF}{data.frame to be filtered}

\item{genelist}{a vector of ID (e.g. Ensembl ID for mitochondria genes) to be excluded from DF}
}
\value{
data.frame without the filtered out genes from the specified list
}
\description{
Exclude the genes (rows) from a data.frame based on a given list
}
