% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-filtering.R
\name{excludeRows}
\alias{excludeRows}
\title{Create a data.frame for which specific rows have been excluded}
\usage{
excludeRows(DF, rowlist)
}
\arguments{
\item{DF}{data.frame}

\item{rowlist}{names of the rows to be excluded}
}
\value{
filtered data.frame
}
\description{
Create a data.frame for which specific rows have been excluded
}
