% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{ExpressedGeneInTissue}
\alias{ExpressedGeneInTissue}
\title{Returns the id (rownames) of the genes that are found in the tissues that have a
breadth of expression equals to or lesser than a given one}
\usage{
ExpressedGeneInTissue(DFa, nbMax = 2, naCol = "nb.tissues",
  comp = FALSE, threshold, omit.zero)
}
\arguments{
\item{DFa}{data.frame; should be logical can be numeric if comp is TRUE}

\item{nbMax}{integer; breadth of expression maximal to consider}

\item{naCol}{character string. Name of the column which gives the breadth of expression of each gene
default: "nb.tissues"}

\item{comp}{logical; should the breadth of expression be computed first from a numeric DFa}

\item{threshold}{numeric, which cutoff of expression has to be used for considering a gene expressed in a tissue or not}

\item{omit.zero}{logical; whether the rows with zeros only should be kept or not}
}
\value{
a list that comprises for each column of the inputted data.frame,
        a vector with the rownames of the genes expressed in a maximal number of tissues (given).
}
\description{
Returns the id (rownames) of the genes that are found in the tissues that have a
breadth of expression equals to or lesser than a given one
}
