% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{extract.HTSc}
\alias{extract.HTSc}
\title{Allows to format older version of irap's output}
\usage{
extract.HTSc(filename, reduce = FALSE, clean = TRUE)
}
\arguments{
\item{filename}{path to the file}

\item{reduce}{boolean. Default: TRUE. Whether the "ID" column should be removed (duplicated with the rownames)}

\item{clean}{boolean. Default:TRUE.}
}
\value{
data.frame of gene expression
}
\description{
Allows to format older version of irap's output
}
