% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{extract.irap}
\alias{extract.irap}
\title{Import data from files produced by an older version of irap.}
\usage{
extract.irap(filename)
}
\arguments{
\item{filename}{character string. Path/filename of a sample (?)}
}
\value{
a data.frame
}
\description{
Import data from files produced by an older version of irap.
}
