% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{extractDendroFromPlotTD}
\alias{extractDendroFromPlotTD}
\title{Extract the dendrogram from the heatmap based on tissueDistance}
\usage{
extractDendroFromPlotTD(DF, method = "ward.D", row = TRUE)
}
\arguments{
\item{DF}{numeric data.frame; output of tissueDistance}

\item{method}{character string, name of the agglomeration method to be used (hclust)
This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single",
"complete", "average" (= UPGMA), "mcquitty" (= WPGMA),
"median" (= WPGMC) or "centroid" (= UPGMC).}

\item{row}{boolean. Default: TRUE. Whether to extract the dendrogram associated to the rows of the heatmap.}
}
\value{
a dendrogram
}
\description{
Extract the dendrogram from the heatmap based on tissueDistance
}
