% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{extractDendroTissueDistance}
\alias{extractDendroTissueDistance}
\title{Create a dendrogram from a tissueDistance data.frame}
\usage{
extractDendroTissueDistance(DF1, trans = TRUE, plot = TRUE, type)
}
\arguments{
\item{DF1}{numeric data.frame; output of tissueDistance}

\item{trans}{boolean. Default: TRUE}

\item{plot}{boolean. Default: TRUE}

\item{type}{a character string specifying the type of phylogeny to be drawn;
Default: "radial"
can also be "phylogram", "cladogram", "fan", "unrooted".}
}
\value{
a dendrogram (phylo object)
}
\description{
Create a dendrogram from a tissueDistance data.frame
}
