% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptors.R
\name{extractSpe}
\alias{extractSpe}
\title{Extract the genes that are specific to a tissue}
\usage{
extractSpe(DF, threshold, strip = TRUE, format = "lg",
  rowNames = TRUE, verbose = FALSE)
}
\arguments{
\item{DF}{numeric data.frame}

\item{threshold}{numeric. Minimal expression to be considered.
For 0, the comparison is strict.}

\item{strip}{boolean. Default: TRUE.}

\item{format}{character string. 'lg' allows to output a long-format data.frame}

\item{rowNames}{boolean. Default: TRUE.}

\item{verbose}{boolean. Default: FALSE.}
}
\value{
a long or wide data.frame
}
\description{
Extract the genes that are specific to a tissue
}
