% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{extractWithinSampleTop3Abundance}
\alias{extractWithinSampleTop3Abundance}
\title{Extract columns that contains "WithInSampleTop3Abundance." in their name}
\usage{
extractWithinSampleTop3Abundance(DF, clean = TRUE)
}
\arguments{
\item{DF}{an expression data.frame}

\item{clean}{boolean; default: TRUE.
Whether rows that contains only null value should be removed}
}
\value{
a data.frame which contains only the columns that contains the tag
}
\description{
Extract columns that contains "WithInSampleTop3Abundance." in their name
}
