% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{extractWithTag}
\alias{extractWithTag}
\title{Allows to create a function that extracts the columns
that contains a specific tag}
\usage{
extractWithTag(tag)
}
\arguments{
\item{tag}{character string that should be look for}
}
\value{
a function
}
\description{
Allows to create a function that extracts the columns
that contains a specific tag
}
