% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{factorDF2string}
\alias{factorDF2string}
\title{Convert all the content of a given data.frame to characters}
\usage{
factorDF2string(DF)
}
\arguments{
\item{DF}{a data.frame}
}
\value{
a character data.frame
}
\description{
Convert all the content of a given data.frame to characters
}
