% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{factoring.rbind}
\alias{factoring.rbind}
\title{From a list of data.frame, add an attribute and collate (row binding) them}
\usage{
factoring.rbind(att1, data, ...)
}
\arguments{
\item{att1}{attribute to add to the data.frames}

\item{data}{list of data.frame}

\item{...}{other arguments for single.factoring2}
}
\value{
a data.frame
}
\description{
From a list of data.frame, add an attribute and collate (row binding) them
}
