% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{factoring}
\alias{factoring}
\title{From a list of filenames, return a list of data.frames after adding an attribute}
\usage{
factoring(att1, data, ...)
}
\arguments{
\item{att1}{an attribute}

\item{data}{filenames}

\item{...}{other arguments for single.factoring}
}
\value{
a list of data.frames
}
\description{
From a list of filenames, return a list of data.frames after adding an attribute
}
