% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-filtering.R
\name{filtre.dataset}
\alias{filtre.dataset}
\title{Keep only the specified rows of a dataframe}
\usage{
filtre.dataset(DF, filter)
}
\arguments{
\item{DF}{data.frame to be filtered}

\item{filter}{vector of the index names of the rows to keep}
}
\value{
data.frame where the rows have been specified
}
\description{
Keep only the specified rows of a dataframe
}
