% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{fixNamesGtex}
\alias{fixNamesGtex}
\title{Homogenise the names of the GTEx data with the other datasets
(in particular with Uhlen et al. )}
\usage{
fixNamesGtex(DF)
}
\arguments{
\item{DF}{a GTEx expression (numeric) data.frame}
}
\value{
a data.frame with homogenised columns names
}
\description{
Homogenise the names of the GTEx data with the other datasets
(in particular with Uhlen et al. )
}
