% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{fixNamesUhlen}
\alias{fixNamesUhlen}
\title{Homogenise the names of the Uhlen et al. data with the other datasets
(in particular with GTEx)}
\usage{
fixNamesUhlen(DF)
}
\arguments{
\item{DF}{a Uhlen expression (numeric) data.frame}
}
\value{
a data.frame with homogenised columns names
}
\description{
Homogenise the names of the Uhlen et al. data with the other datasets
(in particular with GTEx)
}
