% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-extras.R
\name{g_legend}
\alias{g_legend}
\title{Extract the legend from a ggplot}
\usage{
g_legend(a_ggplot)
}
\arguments{
\item{a_ggplot}{a ggplot}
}
\value{
the legend
}
\description{
function developed by Hadley Wickham
found in this (by Michael Kuhn) blog:http://blog.mckuhn.de/2013/04/2d-plot-with-histograms-for-each.html
and was referencing: https://github.com/hadley/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs
}
\author{
Hadley Wickham
}
