% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{gen.add.qual.att}
\alias{gen.add.qual.att}
\title{Create a function that allows to add a qualitative attribute easily}
\usage{
gen.add.qual.att(att)
}
\arguments{
\item{att}{attribute that is fixed in the new function}
}
\value{
a function
}
\description{
Create a function that allows to add a qualitative attribute easily
}
