% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{geneClassify}
\alias{geneClassify}
\title{Classify the genes in function of their expression. Inspired from Uhlén et al. classification.}
\usage{
geneClassify(DF, detect = 0, groupNB, HighThreshold = 10,
  foldLow = 5, foldHigh = 50)
}
\arguments{
\item{DF}{numeric data.frame}

\item{detect}{numeric; default:0. Threshold to consider the gene expressed}

\item{groupNB}{parameter}

\item{HighThreshold}{numeric. Threshold above which the expression is considered high.
default: 10 as it is in Uhlén et al. papers}

\item{foldLow}{numeric. Expression fold to be considered as 'TissueSpecific'. Default:5}

\item{foldHigh}{numeric. Expression fold to be considered as 'TissueSpecific.High'. Default: 50/}
}
\value{
vector of genes names
}
\description{
Classify the genes in function of their expression. Inspired from Uhlén et al. classification.
}
