% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter-plots.R
\name{geneScatters}
\alias{geneScatters}
\title{Scatter plot for the expression of a gene for two datasets across different tissues}
\usage{
geneScatters(ID, DF1, DF2, palette, nameIt = TRUE,
  mapID = barzinePhdData::gene.mapID, logIt = TRUE,
  label.DF1 = "Uhlén et al. (mRNA)\\nlog2(FPKM+1)",
  label.DF2 = "Pandey lab (protein)\\nlog2(PPKM+1)", xcoord, ycoord,
  colNamesDF1 = c("Tissue", "Transcriptomics"),
  colNamesDF2 = c(colNamesDF1[1], "Proteomics"), base_size = 11,
  base_family = "Linux Libertine", label_size = 3.4,
  label_family = base_family, centerTitle = TRUE,
  useCor = "everything")
}
\arguments{
\item{ID}{character string, gene identifier}

\item{DF1}{numeric data.frame or equivalent}

\item{DF2}{numeric data.frame or equivalent}

\item{palette}{named character vector with the colour hex code to use}

\item{nameIt}{logical, default TRUE.
Whether the gene name (from mapID)
should be incorporated to the plot title}

\item{mapID}{named character vector. Map the gene identifiers to their name.
default: \code{\link[barzinePhdData:gene.mapID]{gene.mapID}}}

\item{logIt}{logical, default TRUE. Whether the data should be log2 transform before plotting}

\item{label.DF1}{character string for DF1 labelling. Used as x-axis title}

\item{label.DF2}{character string for DF2 labelling. Used as y-axis title}

\item{xcoord}{vector of two numeric. Gives the min and max of the x-axis}

\item{ycoord}{vector of two numeric. Gives the min and max of the y-axis}

\item{colNamesDF1}{default: c('Tissue','Transcriptomics')}

\item{colNamesDF2}{default: c('Tissue','Proteomics')}

\item{base_size}{numeric, default: 11}

\item{base_family}{default: Linux Libertine}

\item{label_size}{numeric, default 3.4}

\item{label_family}{default: same as base_family}

\item{centerTitle}{logical, default TRUE}

\item{useCor}{default: 'everything'
an optional character string for 'cor'
which gives a method for computing covariances
in the presence of missing values.
This must be (an abbreviation of) one of the strings
"everything", "all.obs", "complete.obs",
"na.or.complete", or "pairwise.complete.obs".}
}
\value{
a plot
}
\description{
Scatter plot for the expression of a gene for two datasets across different tissues
}
