% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-extras.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Create a palette with ggplot2 default colors}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{integer. Number of colours in the palette.}
}
\value{
vector of character strings corresponding to the palette colour hex codes.
}
\description{
Create a palette with ggplot2 default colors
}
\author{
John Colby \url{https://stackoverflow.com/a/8197703/2116422}
}
