% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-related.R
\name{go2genes}
\alias{go2genes}
\title{Takes a list of GO ids and queries all the genes related to it (including the ones for child nodes)}
\usage{
go2genes(goID, annotGO, pathGeneMap, queryList, GOList, ontology,
  exportAnnotGO = TRUE, exportGOList = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{goID}{list of GO ids to retrieve the genes for}

\item{annotGO}{optional; data.frame with information of the go ids and the related genes}

\item{pathGeneMap}{path to the file for the annotGO}

\item{queryList}{argument for createAnnot if no annotGO or pathGeneMap are not provided}

\item{GOList}{GO ids related to the list of the provided GO ids (such as their child nodes)}

\item{ontology}{one from 'BP','CC' or 'MF'}

\item{exportAnnotGO}{logical; whether annotGO should be exported to the environement sharedenv for future use}

\item{exportGOList}{logical; whether GOList should be exported to the environement sharedenv for future use}

\item{verbose}{logical; allows to better track how annotGO and GOList are created/used}

\item{...}{other argument to pass to createGoAnnot}
}
\value{
Gene list associated to the GO ids
}
\description{
Takes a list of GO ids and queries all the genes related to it (including the ones for child nodes)
}
