% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>=\%}
\alias{\%>=\%}
\title{Test more carefully if x is equal or greater than y}
\usage{
x \%>=\% y
}
\arguments{
\item{x}{first member of the comparison}

\item{y}{second member of the comparison}
}
\value{
boolean
}
\description{
Test more carefully if x is equal or greater than y
}
\examples{
8 \%>=\% 8.0
}
