% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGplots.R
\name{greenred}
\alias{greenred}
\title{Generate green-to-red colorscale}
\usage{
greenred(n)
}
\arguments{
\item{n}{Desired number of color elements in the panel.}
}
\value{
Vector of HTML-style RGB colors.
}
\description{
Generate green-to-red colorscale
}
