% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{hampel.test}
\alias{hampel.test}
\title{Apply an Hampel's test on a data.frame}
\usage{
hampel.test(DF, bool = TRUE, threshold = 5.2, ...)
}
\arguments{
\item{DF}{a numeric data.frame}

\item{bool}{logical; default=TRUE. Whether the result should be given as a}

\item{threshold}{numeric. Default: 5.2 (expects a normal distribution)}

\item{...}{additional parameters for stats::mad}
}
\value{
a result data.frame. If logical: TRUE means that the value passes the Hampel test for the given threshold.
                             If numeric, values failing the test are set to 0
}
\description{
Apply an Hampel's test on a data.frame
}
