% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGplots.R
\name{heatmap.2}
\alias{heatmap.2}
\title{Enhanced Heat Map (from gplots)}
\usage{
heatmap.2(x, Rowv = TRUE, Colv = if (symm) "Rowv" else TRUE,
  distfun = dist, hclustfun = hclust, dendrogram = c("both", "row",
  "column", "none"), reorderfun = function(d, w) reorder(d, w),
  symm = FALSE, scale = c("none", "row", "column"), na.rm = TRUE,
  revC = identical(Colv, "Rowv"), add.expr, breaks, symbreaks = any(x <
  0, na.rm = TRUE) || scale != "none", col = "heat.colors", colsep,
  rowsep, sepcolor = "white", sepwidth = c(0.05, 0.05), cellnote,
  notecex = 1, notecol = "cyan", na.color = par("bg"),
  trace = c("column", "row", "both", "none"), tracecol = "cyan",
  hline = median(breaks), vline = median(breaks), linecol = tracecol,
  margins = c(5, 5), ColSideColors, RowSideColors, cexRow = 0.2 +
  1/log10(nr), cexCol = 0.2 + 1/log10(nc), labRow = NULL,
  labCol = NULL, srtRow = NULL, srtCol = NULL, adjRow = c(0, NA),
  adjCol = c(NA, 0), offsetRow = 0.5, offsetCol = 0.5,
  colRow = NULL, colCol = NULL, key = TRUE, keysize = 1.5,
  density.info = c("histogram", "density", "none"), denscol = tracecol,
  symkey = any(x < 0, na.rm = TRUE) || symbreaks, densadj = 0.25,
  key.title = NULL, key.xlab = NULL, key.ylab = NULL,
  key.xtickfun = NULL, key.ytickfun = NULL, key.par = list(),
  main = NULL, xlab = NULL, ylab = NULL, lmat = NULL,
  lhei = NULL, lwid = NULL, extrafun = NULL, ...)
}
\arguments{
\item{x}{numeric matrix of the values to be plotted.}

\item{Rowv}{determines if and how the row dendrogram should be reordered.    By default, it is TRUE, which implies dendrogram is computed and reordered based on row means. If NULL or FALSE, then no dendrogram is computed and no reordering is done. If a dendrogram, then it is used "as-is", ie without any reordering. If a vector of integers, then dendrogram is computed and reordered based on the order of the vector.}

\item{Colv}{determines if and how the column dendrogram should be reordered.    Has the options as the Rowv argument above and additionally when x is a square matrix, Colv="Rowv" means that columns should be treated identically to the rows.}

\item{distfun}{function used to compute the distance (dissimilarity) between both rows and columns. Defaults to dist.}

\item{hclustfun}{function used to compute the hierarchical clustering when Rowv or Colv are not dendrograms. Defaults to hclust.}

\item{dendrogram}{character string indicating whether to draw 'none', 'row', 'column' or 'both' dendrograms. Defaults to 'both'. However, if Rowv (or Colv) is FALSE or NULL and dendrogram is 'both', then a warning is issued and Rowv (or Colv) arguments are honoured.}

\item{reorderfun}{function(d, w) of dendrogram and weights for reordering the row and column dendrograms. The default uses stats{reorder.dendrogram}}

\item{symm}{logical indicating if x should be treated symmetrically; can only be true when x is a square matrix.}

\item{scale}{character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. The default is "none".}

\item{na.rm}{logical indicating whether NA's should be removed.}

\item{revC}{logical indicating if the column order should be reversed for plotting, such that e.g., for the symmetric case, the symmetry axis is as usual.}

\item{add.expr}{expression that will be evaluated after the call to image. Can be used to add components to the plot.}

\item{breaks}{(optional) Either a numeric vector indicating the splitting points for binning x into colors, or a integer number of break points to be used, in which case the break points will be spaced equally between min(x) and max(x).}

\item{symbreaks}{Boolean indicating whether breaks should be made symmetric about 0. Defaults to TRUE if the data includes negative values, and to FALSE otherwise.}

\item{col}{colors used for the image. Defaults to heat colors (heat.colors).}

\item{colsep}{(optional) vector of integers indicating which columns or rows should be separated from the preceding columns or rows by a narrow space of color sepcolor.}

\item{rowsep}{(optional) vector of integers indicating which columns or rows should be separated from the preceding columns or rows by a narrow space of color sepcolor.}

\item{sepcolor}{(optional) vector of integers indicating which columns or rows should be separated from the preceding columns or rows by a narrow space of color sepcolor.}

\item{sepwidth}{(optional) Vector of length 2 giving the width (colsep) or height (rowsep) the separator box drawn by colsep and rowsep as a function of the width (colsep) or height (rowsep) of a cell. Defaults to c(0.05, 0.05)}

\item{cellnote}{(optional) matrix of character strings which will be placed within each color cell, e.g. p-value symbols.}

\item{notecex}{(optional) numeric scaling factor for cellnote items.}

\item{notecol}{(optional) character string specifying the color for cellnote text. Defaults to "cyan". Can be a matrix to personalise each cell.}

\item{na.color}{Color to use for missing value (NA). Defaults to the plot background color.}

\item{trace}{character string indicating whether a solid "trace" line should be drawn across 'row's or down 'column's, 'both' or 'none'. The distance of the line from the center of each color-cell is proportional to the size of the measurement. Defaults to 'column'.}

\item{tracecol}{character string giving the color for "trace" line. Defaults to "cyan".}

\item{hline}{Vector of values within cells where a horizontal or vertical dotted line should be drawn. The color of the line is controlled by linecol. Horizontal lines are only plotted if trace is 'row' or 'both'. Vertical lines are only drawn if trace 'column' or 'both'. hline and vline default to the median of the breaks, linecol defaults to the value of tracecol.}

\item{vline}{Vector of values within cells where a horizontal or vertical dotted line should be drawn. The color of the line is controlled by linecol. Horizontal lines are only plotted if trace is 'row' or 'both'. Vertical lines are only drawn if trace 'column' or 'both'. hline and vline default to the median of the breaks, linecol defaults to the value of tracecol.}

\item{linecol}{Vector of values within cells where a horizontal or vertical dotted line should be drawn. The color of the line is controlled by linecol. Horizontal lines are only plotted if trace is 'row' or 'both'. Vertical lines are only drawn if trace 'column' or 'both'. hline and vline default to the median of the breaks, linecol defaults to the value of tracecol.}

\item{margins}{numeric vector of length 2 containing the margins (see par(mar= *)) for column and row names, respectively.}

\item{ColSideColors}{(optional) character vector of length ncol(x) containing the color names for a horizontal side bar that may be used to annotate the columns of x.}

\item{RowSideColors}{(optional) character vector of length nrow(x) containing the color names for a vertical side bar that may be used to annotate the rows of x.}

\item{cexRow}{positive numbers, used as cex.axis in for the row or column axis labeling. The defaults currently only use number of rows or columns, respectively.}

\item{cexCol}{positive numbers, used as cex.axis in for the row or column axis labeling. The defaults currently only use number of rows or columns, respectively.}

\item{labRow}{character vectors with row and column labels to use; these default to rownames(x) or colnames(x), respectively.}

\item{labCol}{character vectors with row and column labels to use; these default to rownames(x) or colnames(x), respectively.}

\item{srtRow}{angle of row/column labels, in degrees from horizontal}

\item{srtCol}{angle of row/column labels, in degrees from horizontal}

\item{adjRow}{2-element vector giving the (left-right, top-bottom) justification of row/column labels (relative to the text orientation).}

\item{adjCol}{2-element vector giving the (left-right, top-bottom) justification of row/column labels (relative to the text orientation).}

\item{offsetRow}{Number of character-width spaces to place between row/column labels and the edge of the plotting region.}

\item{offsetCol}{Number of character-width spaces to place between row/column labels and the edge of the plotting region.}

\item{colRow}{color of row/column labels, either a scalar to set the color of all labels the same, or a vector providing the colors of each label item}

\item{colCol}{color of row/column labels, either a scalar to set the color of all labels the same, or a vector providing the colors of each label item}

\item{key}{logical indicating whether a color-key should be shown.}

\item{keysize}{numeric value indicating the size of the key}

\item{density.info}{character string indicating whether to superimpose a 'histogram', a 'density' plot, or no plot ('none') on the color-key.}

\item{denscol}{character string giving the color for the density display specified by density.info, defaults to the same value as tracecol.}

\item{symkey}{Boolean indicating whether the color key should be made symmetric about 0. Defaults to TRUE if the data includes negative values, and to FALSE otherwise.}

\item{densadj}{Numeric scaling value for tuning the kernel width when a density plot is drawn on the color key. (See the adjust parameter for the density function for details.) Defaults to 0.25.}

\item{key.title}{main title of the color key. If set to NA no title will be plotted.}

\item{key.xlab}{x axis label of the color key. If set to NA no label will be plotted.}

\item{key.ylab}{y axis label of the color key. If set to NA no label will be plotted.}

\item{key.xtickfun}{function computing tick location and labels for the xaxis of the color key. Returns a named list containing parameters that can be passed to axis. See examples.}

\item{key.ytickfun}{function computing tick location and labels for the y axis of the color key. Returns a named list containing parameters that can be passed to axis. See examples.}

\item{key.par}{graphical parameters for the color key. Named list that can be passed to par.}

\item{main}{main, x- and y-axis titles; defaults to none.}

\item{xlab}{main, x- and y-axis titles; defaults to none.}

\item{ylab}{main, x- and y-axis titles; defaults to none.}

\item{lmat}{visual layout: position matrix, column height, column width. See below for details}

\item{lhei}{visual layout: position matrix, column height, column width. See below for details}

\item{lwid}{visual layout: position matrix, column height, column width. See below for details}

\item{extrafun}{A function to be called after all other work. See examples.}

\item{...}{additional arguments passed on to image}
}
\value{
invisibly, a list with components
rowInd	row index permutation vector as returned by order.dendrogram.
colInd	column index permutation vector.
call the matched call
rowMeans, rowSDs
mean and standard deviation of each row: only present if scale="row"
colMeans, colSDs
mean and standard deviation of each column: only present if scale="column"
carpet	reordered and scaled 'x' values used generate the main 'carpet'
rowDendrogram	row dendrogram, if present
colDendrogram	column dendrogram, if present
breaks	values used for color break points
col colors used
vline center-line value used for column trace, present only if trace="both" or trace="column"
hline	center-line value used for row trace, present only if trace="both" or trace="row"
colorTable	 A three-column data frame providing the lower and upper bound and color for each bin
layout	A named list containing the values used for lmat, lhei, and lwid.
}
\description{
A heat map is a false color image (basically image(t(x))) with a dendrogram added to the left side and/or to the top. Typically, reordering of the rows and columns according to some set of values (row or column means) within the restrictions imposed by the dendrogram is carried out.
This heatmap provides a number of extensions to the standard R heatmap function.
}
