% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-wrappers.R
\name{heatmapAlpabet}
\alias{heatmapAlpabet}
\title{Draw a customised heatmap where the different variables are sorted alphabetically}
\usage{
heatmapAlpabet(data, label.A, label.B, out.A, out.B, digits, key,
  col = barzinePhdR::colCbc(), sorting = TRUE, margins, ...,
  report = TRUE, index = "####", annot = "diag")
}
\arguments{
\item{data}{correlation (numeric) data.frame to plot}

\item{label.A}{character string. Name of the first data.frame used for the correlation matrix
for the additional text.}

\item{label.B}{character string. Name of the second data.frame used for the correlation matrix
for the additional text.}

\item{out.A}{character string for the plot (x-axis).
Name of the first data.frame used for the correlation matrix}

\item{out.B}{character string for the plot (y-axis).
Name of the second data.frame used for the correlation matrix}

\item{digits}{integer. default: 1. How many digits should be kept for the annotation.}

\item{key}{boolean. default: TRUE. Whether the key of the heatmap should also be drawn.}

\item{col}{color palette to use; default barzinePhdR::colC(),
i.e. grDevices::colorRampPalette('aliceblue','darkcyan')}

\item{sorting}{boolean; default: TRUE.
Whether the input data.frame should be rordered alphabetically before being plotted.}

\item{margins}{numeric vector of length 2 containing the margins
(see par(mar= *)) for column and row names, respectively.}

\item{...}{other parameters that can be handled by (gplots) heatmap.2}

\item{report}{boolean. Default: TRUE. Output additional text (example for an (r)html report)}

\item{index}{character string. default:"####" Allows to customise the level of indexing through markdown tags.}

\item{annot}{character string. Default: 'diag'. Annotate only the diagonal of the heatmap
with the value of correlation.}
}
\value{
a heatmap plot
}
\description{
Draw a customised heatmap where the different variables are sorted alphabetically
}
