% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-wrappers.R
\name{heatmapWithClustering}
\alias{heatmapWithClustering}
\title{Draw a customised heatmap with marked clusters in the margins for a correlation matrix}
\usage{
heatmapWithClustering(data, label.A, label.B, out.A, out.B,
  dend = "none", margins, col = barzinePhdR::colCbc(), ...,
  report = TRUE, index = "####")
}
\arguments{
\item{data}{correlation data.frame to plot}

\item{label.A}{character string. Name of the first data.frame used for the correlation matrix
for the additional text.}

\item{label.B}{character string. Name of the second data.frame used for the correlation matrix
for the additional text.}

\item{out.A}{character string for the plot (x-axis).
Name of the first data.frame used for the correlation matrix}

\item{out.B}{character string for the plot (y-axis).
Name of the second data.frame used for the correlation matrix}

\item{dend}{character string that correspond to \code{\link[gplots]{heatmap.2}} "dendrogram"
indicating whether to draw 'none', 'row', 'column' or 'both' dendrograms.
Defaults to 'both'. However, if Rowv (or Colv) is FALSE or NULL and dendrogram is 'both',
then a warning is issued and Rowv (or Colv) arguments are honoured.}

\item{margins}{numeric vector of length 2 containing the margins
(see par(mar= *)) for column and row names, respectively.}

\item{col}{color palette to use; default barzinePhdR::colC(),
i.e. grDevices::colorRampPalette('aliceblue','darkcyan')}

\item{...}{other parameters that can be handled by \code{\link[gplots]{heatmap.2}}}

\item{report}{boolean. Default: TRUE. Output additional text (example for an (r)html report)}

\item{index}{character string. default:"####" Allows to customise the level of indexing through markdown tags.}
}
\value{
a heatmap plot
}
\description{
Draw a customised heatmap with marked clusters in the margins for a correlation matrix
}
