% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barzinePhdR-package.R
\name{initialise}
\alias{initialise}
\title{barzinePhdR}
\usage{
initialise()
}
\description{
This package contains all the require R functions supporting the analyses
in [https://github.com/barzine/phd-analyses](https://github.com/barzine/phd-analyses)
which are presented in [https://github.com/barzine/thesis](https://github.com/barzine/thesis)
}
