% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{internalJaccardIndMatrix}
\alias{internalJaccardIndMatrix}
\title{Create matrix of jaccard indices based on the output of ExpressedGeneInTissue}
\usage{
internalJaccardIndMatrix(DFa, geneList, ...)
}
\arguments{
\item{DFa}{data.frame}

\item{geneList}{list of the genes found in tissue}

\item{...}{other parameters for ExpressedGeneInTissue}
}
\value{
a matrix
}
\description{
Create matrix of jaccard indices based on the output of ExpressedGeneInTissue
}
