% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Intersect}
\alias{Intersect}
\title{Allow to intersect more than two sets at a time}
\usage{
Intersect(...)
}
\arguments{
\item{...}{comma-separetated list of sets}
}
\value{
the intersection of all the sets
}
\description{
Allow to intersect more than two sets at a time
}
