% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{jaccardInd}
\alias{jaccardInd}
\title{Compute Jaccard index}
\usage{
jaccardInd(list1, list2, universeSize, pvalue = FALSE, indexSignif = 2,
  pvalSig = 3, figure = FALSE, Col = c("coral4", "darkorchid1"),
  categories, condition, policeFont = "Linux Libertine", saveFile = NA,
  outputType = "index", ...)
}
\arguments{
\item{list1}{vector of gene names from a first element to compare}

\item{list2}{vector of genes names from a second element to compare}

\item{universeSize}{positive integer. Number of elements in the universe}

\item{pvalue}{boolean. Default: FALSE. Whether a p-value should be computed}

\item{indexSignif}{positive integer. Default: 2. Number of significant digits to add for the Jaccard index on the figure.}

\item{pvalSig}{positive integer. Default: 3. Number of significant digits to add for the p-value on the figure.}

\item{figure}{boolean. default: FALSE. Whether a figure showing the overlap should be displayed}

\item{Col}{Colour palette: vector of two character strings. Colours to use to draw the two elements being compared in the venn diagram}

\item{categories}{vector of two character strings that are the names of the two elements being compared.}

\item{condition}{character string. Name of the condition compared between the two elements.}

\item{policeFont}{character string. Name of the font to use.}

\item{saveFile}{path (character string). Default: NA. Allows to save the figure (if figure=TRUE) at the given path.}

\item{outputType}{character string. Whether "index" to return the Jaccard index;
"pval" for the p-value associated or
"list" for a list comprising the index and the pvalue,
"intersect" intersection of the two lists.}

\item{...}{other parameters for grDevices::cairo_pdf}
}
\value{
depends on outputType. Refer to that argument for the possible output.
}
\description{
Compute Jaccard index
}
