% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{lg.format}
\alias{lg.format}
\title{Format a data.frame in long format and annotate columns}
\usage{
lg.format(DF, RowNames = TRUE, TRow = "Gene", categoriesA, categoriesB,
  nameA = "condition", nameB = "study", col1 = "variable")
}
\arguments{
\item{DF}{a data.frame}

\item{RowNames}{weither the rownames should be included as new column}

\item{TRow}{Name of the column for the row.names. Default="Gene".}

\item{categoriesA}{vector comprising the different string that should be look into
the column names of the input data.frame}

\item{categoriesB}{vector comprising the different string that should be look into
the column names of the input data.frame}

\item{nameA}{Name of the new column to add for categoriesA in the result data.frame}

\item{nameB}{Name of the new column to add for categoriesB in the result data.frame}

\item{col1}{the name of the column to parse for the categories.
default: "variable" (default name given by reshape2::melt for expression data.frame )}
}
\value{
a long data.frame
}
\description{
Format a data.frame in long format and annotate columns
}
