% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{listFromFile}
\alias{listFromFile}
\title{Create a vector from a file where the data is formatted as a list
with each element on a newline}
\usage{
listFromFile(filename)
}
\arguments{
\item{filename}{path to the file}
}
\value{
a vector
}
\description{
Create a vector from a file where the data is formatted as a list
with each element on a newline
}
