% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{listFromFileBiomart}
\alias{listFromFileBiomart}
\title{Create a vector from a file downloaded from Biomart
where the data is formatted as a list
each newline comprises a gene id and transcript id.}
\usage{
listFromFileBiomart(filename, out = TRUE)
}
\arguments{
\item{filename}{path to the file}

\item{out}{boolean. Default:TRUE. Whether the genes ID should be returned}
}
\value{
a vector of gene ID or a data.frame with the genes and transcripts ID
}
\description{
Create a vector from a file downloaded from Biomart
where the data is formatted as a list
each newline comprises a gene id and transcript id.
}
