% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{listgenes}
\alias{listgenes}
\title{Retrieve all the genes ID from a file}
\usage{
listgenes(filename, h = "Gene.ID")
}
\arguments{
\item{filename}{path to the file}

\item{h}{character string. Header of the files. Default: "Gene.ID"}
}
\value{
a vector or a named vector if there are headers
}
\description{
Retrieve all the genes ID from a file
}
