% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{ListOfExpressedGenes}
\alias{ListOfExpressedGenes}
\title{Give the list of expressed genes in a study in function of the selected type of comparison}
\usage{
ListOfExpressedGenes(DF, threshold = 0, out = "rownames",
  type = "gt", nb)
}
\arguments{
\item{DF}{numeric data.frame}

\item{threshold}{numeric, default: 0. Threshold above which a gene is considered as expressed}

\item{out}{result format;
'rownames': names of the relevant genes
'data.frame': data.frame with the relevant genes
'df': alias of 'data.frame'
'index': indices of the rows of the relevant genes
'ind': alias of 'index'}

\item{type}{type of comparison to be considered.
'gt': expression should be strictly greater than threshold.
'ge': expression should be greater than or equal to threshold.
'eq': expression should be equal to threshold.
'lt': expression should strictly be lesser than threshold
'le': expression should be lesser than or equal to threshold
'notDetected': No expression is detected above or at threshold
'ubi': expression is detected everywhere
'unique': expression is detected solely in one condition
'mixedH': expression is detected above threshold in 'nb' tissues
'mixedl': expression is detected below threshold in 'nb' tissues}

\item{nb}{Number of tissues that one should be expressed to be considered for mixed expression}
}
\value{
several types of objects are available. Selection through "out"
}
\description{
Give the list of expressed genes in a study in function of the selected type of comparison
}
