% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{Load}
\alias{Load}
\title{Load function that returns the object(s) instead of attaching it to the global namespace
stolen from K. Rudolph and M. Schubert in the `ebits` repo}
\usage{
Load(filename)
}
\arguments{
\item{filename}{filename to be loaded in the global environment}
}
\value{
the object content which needs to be stored for later use
}
\description{
Load function that returns the object(s) instead of attaching it to the global namespace
stolen from K. Rudolph and M. Schubert in the `ebits` repo
}
