% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{load_and_check_3DF}
\alias{load_and_check_3DF}
\title{Load three data.frames from file paths and check that they have the same dimensions}
\usage{
load_and_check_3DF(path1, path2, path3)
}
\arguments{
\item{path1}{file path to the first data.frame}

\item{path2}{file path to the second data.frame}

\item{path3}{file path to the third data.frame}
}
\value{
a list of three data.frames
}
\description{
Load three data.frames from file paths and check that they have the same dimensions
}
