% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log2.na}
\alias{log2.na}
\title{The correct answer -Inf for log2(0) is changed to NA}
\usage{
log2.na(x)
}
\arguments{
\item{x}{numeric value}
}
\value{
log2(x) or NA for log2(O)
}
\description{
The correct answer -Inf for log2(0) is changed to NA
}
