% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{make.2ldf}
\alias{make.2ldf}
\title{Create a data.frame from the elements of two vectors that have the same names}
\usage{
make.2ldf(a, b, colN)
}
\arguments{
\item{a}{first vector}

\item{b}{second vector}

\item{colN}{Colnames for the new data.frame}
}
\value{
a data.frame with two colums
}
\description{
Create a data.frame from the elements of two vectors that have the same names
}
