% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{make.3ldf}
\alias{make.3ldf}
\title{Create a data.frame from the elements of three vectors that have the same names}
\usage{
make.3ldf(a, b, c, colN)
}
\arguments{
\item{a}{first vector}

\item{b}{second vector}

\item{c}{third vector}

\item{colN}{Colnames for the new data.frame}
}
\value{
a data.frame with three columns
}
\description{
Create a data.frame from the elements of three vectors that have the same names
}
