% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-wrappers.R
\name{matchColCond}
\alias{matchColCond}
\title{Allows to display the correct colour annotation}
\usage{
matchColCond(Names, cond, palette)
}
\arguments{
\item{Names}{vector of character strings. name of the columns}

\item{cond}{vector of character strings. Name of the tissues/conditions to map the columns}

\item{palette}{named vector palette. (names are the same as condition)}
}
\value{
a properly formatted vector of colours
}
\description{
Allows to display the correct colour annotation
}
