% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{matrix.overlap_res}
\alias{matrix.overlap_res}
\title{Calculate the overlap of specific genes (uniquely expressed in one tissue)
between two studies.}
\usage{
matrix.overlap_res(DF1, DF2, ratio = -1, common = FALSE,
  selectSpe = FALSE, verbose = TRUE, report = "html")
}
\arguments{
\item{DF1}{Expression data.frame of the first studie to be compared}

\item{DF2}{Expression data.frame of the second studie to be compared}

\item{ratio}{numeric; multiplicator for the number of genes used
for the second data.frame compared to the first one.
"-1" if all specific genes of each data.frame should be compared}

\item{common}{argument passed on to overlapProtTrans.Specific; default: FALSE}

\item{selectSpe}{boolean; default: FALSE. Allows to select the specific genes in the data.frames
if not done beforehand.}

\item{verbose}{boolean; default: TRUE. Allows outputting additional information}

\item{report}{character string. Default: 'html'. Other option: ""
When report='html' use cat.html instead of print}
}
\value{
a list of two matrices. One matrix with the overlap (in percentage)
        and another matrix with the corresponding p-values.
}
\description{
Calculate the overlap of specific genes (uniquely expressed in one tissue)
between two studies.
}
