% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{merge_biol_rep}
\alias{merge_biol_rep}
\title{Merge the biological replicates in a data.frame together}
\usage{
merge_biol_rep(annotDF, DF, FUN = "rowMeans", alone = FALSE, ...)
}
\arguments{
\item{annotDF}{annotation dataframe that indicates the organisation of the biological replicates}

\item{DF}{data.frame}

\item{FUN}{the function with which the replicates biological are merged. Default: "rowMeans"}

\item{alone}{if all the columns are replicates of a single sample; default=FALSE}

\item{...}{not implemented}
}
\value{
a data.frame
}
\description{
Merge the biological replicates in a data.frame together
}
