% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{merge_replicates}
\alias{merge_replicates}
\title{Merge the replicates of a data.frame together}
\usage{
merge_replicates(annot.DF, DF, group = "double", FUN = "rowMeans",
  FUNT, FUNB, ...)
}
\arguments{
\item{annot.DF}{annotation dataframe that indicates
the organisation of the biological and technical replicates}

\item{DF}{data.frame}

\item{group}{"double" (default); "Biological" ("b" or "B"); "Technical" ("t" or "T")}

\item{FUN}{the function with which the replicates (biological and/or technical) are merged. Default: "rowMeans"}

\item{FUNT}{function with which the technical replicates are merged}

\item{FUNB}{function with which the biological replicates are merged}

\item{...}{not implemented}
}
\value{
data.frame
}
\description{
Merge the replicates of a data.frame together
}
