% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{merge_tech_rep}
\alias{merge_tech_rep}
\title{Merge the techincal replicates in a data.frame together}
\usage{
merge_tech_rep(annotDF, DF, FUN = "rowMeans")
}
\arguments{
\item{annotDF}{annotation dataframe that indicates the organisation of the technical replicates}

\item{DF}{data.frame}

\item{FUN}{the function with which the replicates technical are merged. Default: "rowMeans"}
}
\value{
a data.frame
}
\description{
Merge the techincal replicates in a data.frame together
}
