% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-wrappers.R
\name{mHeatmap}
\alias{mHeatmap}
\title{Wrapper around heatmap.2; allows to annotate with the colours on the side.}
\usage{
mHeatmap(DF, method = "pearson", use = "everything",
  hclustMethod = "ward.D", trace = "none", dendrogram = "row", col,
  cexRow = 1, cexCol = 1, annotate = FALSE, signifdigit = 1,
  rowsidecolors, colsidecolors, density.info, notecol = "black",
  ColsideMode = "match", RowsideMode = "extract", srtCol = 45,
  margins = c(12, 8), common.cond, datasetCol, TissueCol, baseFont, ...,
  parenthesis = TRUE, key.title)
}
\arguments{
\item{DF}{data.frame or matrix to use for the heatmap}

\item{method}{string to pick the method with which to compute the correlation.
One of "pearson" (default), "kendall", or "spearman"}

\item{use}{character string giving a method for computing covariances in the presence of missing values.
This must be (an abbreviation of) one of the strings
"everything" (default), "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs".}

\item{hclustMethod}{character string. he agglomeration method to be used.
This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{trace}{character string indicating whether a solid "trace" line should be drawn across 'row's or down 'column's, 'both' or 'none'. The distance of the line from the center of each color-cell is proportional to the size of the measurement. Defaults to 'none'.}

\item{dendrogram}{character string indicating whether to draw 'none', 'row', 'column' or 'both' dendrograms. Defaults to 'both'. However, if Rowv (or Colv) is FALSE or NULL and dendrogram is 'both', then a warning is issued and Rowv (or Colv) arguments are honoured.}

\item{col}{colors used for the image. Defaults to col=colorRampPalette(c('ghostwhite','darkcyan'))}

\item{cexRow, cexCol}{positive numbers, used as cex.axis in for the row or column axis labeling. The defaults currently only use number of rows or columns, respectively.}

\item{annotate}{boolean. Whether to add the correlations as annotation on the heatmap}

\item{signifdigit}{integer. Number of digit to show for the correlation showed as annotation. Default:1}

\item{rowsidecolors}{(optional) character vector of length nrow(x) containing the color names for a vertical side bar that may be used to annotate the rows of x.
Otherwise can be created internally}

\item{colsidecolors}{(optional) character vector of length ncol(x) containing the color names for a horizontal side bar that may be used to annotate the columns of x.
Otherwise can be created internally}

\item{density.info}{character string indicating whether to superimpose a 'histogram', a 'density' plot, or no plot ('none') on the color-key.}

\item{notecol}{character string. Colour for the annotation}

\item{ColsideMode}{character string that allows to create the annotation colours based on the column names
one of the following strings: 'match', 'extract' or 'extractSpace'}

\item{RowsideMode}{character string that allows to create the annotation colours based on the row names (of the correlation matrix)
one of the following strings: 'extract', 'extractSpace', 'extract2' or 'extractSpace2'}

\item{srtCol}{angle of column labels, in degrees from horizontal. Default: 45}

\item{margins}{numeric vector of length 2 containing the margins (see par(mar= *)) for column and row names, respectively. Default: c(8,12)}

\item{common.cond}{argument for matchColCond}

\item{datasetCol}{colour palette for the datasets in the form of a named vector}

\item{TissueCol}{colour palette for the tissues in the form of a named vector}

\item{baseFont}{character string. Allows to pick the font of the figure}

\item{...}{other parameters that can be used by heatmap.2}

\item{parenthesis}{boolean. Whether the column names comprises parenthesis}

\item{key.title}{main title of the color key. If set to NA no title will be plotted.}
}
\value{
a heatmap
}
\description{
Wrapper around heatmap.2; allows to annotate with the colours on the side.
}
