% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{mixed.low}
\alias{mixed.low}
\title{Provide the list of mixed low genes}
\usage{
mixed.low(DF, threshold = 10, detect = 1)
}
\arguments{
\item{DF}{a numeric data.frame}

\item{threshold}{Maximal level of expression when detected; default: 10.0}

\item{detect}{numeric. Minimal level to be considered as detected; default: 1}
}
\value{
a named vector with the number of samples the genes has been detected as mixed low
}
\description{
For each gene (rows) the number of samples/tissues (columns)
in which they are expressed is given as the content of the named vector.
Genes have to be consistantly lowly expressed (ie expressed lower a given threshold)
when they are detected (ie genes have to be expressed above another given threshold).
}
