% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{mixed.low2}
\alias{mixed.low2}
\title{Provide the list of genes that have a low expression when they are expressed}
\usage{
mixed.low2(DF, threshold = 10, detect = 1, clean = TRUE)
}
\arguments{
\item{DF}{a numeric data.frame}

\item{threshold}{Maximal level of expression when detected; default: 10.0}

\item{detect}{numeric. Minimal level to be considered as detected; default: 1}

\item{clean}{logical; default: TRUE. Wheter the genes that aren't mixed low
should be removed from the final result.}
}
\value{
named vector. Names of the genes that have a mixed low expression.
}
\description{
Provide the list of genes that have a low expression when they are expressed
}
