% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{mkMatrix}
\alias{mkMatrix}
\title{Take two data.frames to create one and can log2 the content at once
while changing the -Inf to NA}
\usage{
mkMatrix(x, y, logT = FALSE)
}
\arguments{
\item{x}{a numeric data.frame}

\item{y}{a numeric data.frame}

\item{logT}{logical; default: FALSE}
}
\value{
one numeric data.frame
}
\description{
Take two data.frames to create one and can log2 the content at once
while changing the -Inf to NA
}
