% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{multi.add.qual.att}
\alias{multi.add.qual.att}
\title{Create a dataframe filled with a value with the same number of rows than a given df
and then bind the two together (new data.frame as new columns)}
\usage{
multi.add.qual.att(df, att, val, ...)
}
\arguments{
\item{df}{data.frame of reference}

\item{att}{vector of character strings}

\item{val}{vector of values to use to fill the new columns}

\item{...}{additionnal arguments for base::data.frame}
}
\value{
a data.frame
}
\description{
Create a dataframe filled with a value with the same number of rows than a given df
and then bind the two together (new data.frame as new columns)
}
