% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{multibarplotsDiversityCond}
\alias{multibarplotsDiversityCond}
\title{For easier comparison, plot the uniquely expressed genes (colored by tissues) in two studies
Allows to consider several thresholds of expression for the second study.}
\usage{
multibarplotsDiversityCond(DF1, DF2, threshold1 = 0, threshold2 = c(0,
  1, 5), label1 = "Proteins", midLabel1 = "(", endLabel1 = ")",
  label2 = "mRNA", midLabel2 = "(", endLabel2 = "FPKM)",
  sorted = TRUE, common = TRUE, colorpal = NULL, publish = TRUE,
  output = "count", verbose = TRUE, decreasing = TRUE, ...)
}
\arguments{
\item{DF1}{numeric data.frame (first study expression data)}

\item{DF2}{numeric data.frame (second study expression data)}

\item{threshold1}{numeric. Expression above which a gene is considered as expressed for the first study}

\item{threshold2}{a vector of numeric. Expression values above which a gene is considered as expressed for the second study}

\item{label1}{character string. Label for the first study to use on the plot (first part)
Default: 'Proteins'}

\item{midLabel1}{character string. Label for the first study to use on the plot (first part before threshold)
Default: '('}

\item{endLabel1}{character string. Label for the first study to use on the plot (second part after threshold)
Default: ')'}

\item{label2}{character string. Label for the second study to use on the plot (first part)
Default: 'mRNA'}

\item{midLabel2}{character string. Label for the second study to use on the plot (first part before threshold)
Default: '('}

\item{endLabel2}{character string. Label for the second study to use on the plot (second part after threshold)
Default: 'FPKM)'}

\item{sorted}{boolean. Default: TRUE. Whether the tissues should be sorted in function of their number of tissue specific genes}

\item{common}{boolean. Default: TRUE. Whether the two studies should share identical rownames and colnames}

\item{colorpal}{colour palette to use in the figure (done with ggplot2::scale_fill_manual)}

\item{publish}{boolean. Default: TRUE. Whether to apply ggplot2::theme_bw to the plot.}

\item{output}{character string. Switch that allows to choose between 'count' for the count of unique genes across the tissues
or a ratio based on the distribution of the tissue specific genes across each study.}

\item{verbose}{boolean. Default: TRUE.}

\item{decreasing}{boolean. Default: TRUE. Whether the sorting should be done in decreasing order.}

\item{...}{other arguments that can be used by ggplot2::theme_bw()}
}
\value{
a list of object that would allow to create a figure
}
\description{
For easier comparison, plot the uniquely expressed genes (colored by tissues) in two studies
Allows to consider several thresholds of expression for the second study.
}
