% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-related.R
\name{nameID}
\alias{nameID}
\title{Retrieve gene symbols (hgnc) for Ensembl gene IDs with biomaRt}
\usage{
nameID(geneID, out = "nameOnly", ens)
}
\arguments{
\item{geneID}{vector of Ensembl gene ids}

\item{out}{'all' for a named vector where the gene symbol are the content
     and the Ensembl ids are the names
'DF' for a data.frame where the first column contains the Ensembl ids
and the second one the gene symbols.
'nameOnly' for the gene symbols only}

\item{ens}{Formal class 'Mart' from package biomaRt}
}
\value{
gene symbols (with or without their corresponding Ensembl ids)
}
\description{
function which has generated "gene.mapID"
}
