% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptors.R
\name{nbGenesPerCond}
\alias{nbGenesPerCond}
\title{Present the number of genes that are expressed in each tissue}
\usage{
nbGenesPerCond(DF, reorder = TRUE, xlab, ylab, title, Y, thm,
  publish = TRUE)
}
\arguments{
\item{DF}{numeric data.frame that contains expresion data}

\item{reorder}{boolean. Default: TRUE. Whether the tissues
should be sorted in decreasing order of expressed genes}

\item{xlab}{character string. Label for the x-axis}

\item{ylab}{character string. Label for the y-axis}

\item{title}{character string. Title on the figure}

\item{Y}{a length-2 numeric vector. Limites of the y-axis}

\item{thm}{function theme( ) with the list of additional elements to apply}

\item{publish}{boolean. Default: TRUE.}
}
\value{
a figure
}
\description{
Present the number of genes that are expressed in each tissue
}
