% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptors.R
\name{nbGenesPerCond.unique}
\alias{nbGenesPerCond.unique}
\title{Plot the distribution of genes across the columns.
Genes found uniquely in one condition are highlighted in another colour.}
\usage{
nbGenesPerCond.unique(DF, reorder = TRUE, xlab = "Tissue", ylab, title,
  legend.title = "Gene is", publish = TRUE, threshold = 0, thm, Y,
  remove.legend = FALSE, legendPosition = "bottom", out = "DF",
  print = TRUE)
}
\arguments{
\item{DF}{numerical data.frame. Rownames have to be meaningful.}

\item{reorder}{boolean; default: TRUE.
Whether the columns should be reorder by the number of expressed genes instead of their names.}

\item{xlab}{character string; used as x-axis label.}

\item{ylab}{character string; used as y-axis label.}

\item{title}{character string; title of the plot.}

\item{legend.title}{character string; what to use for the colour legend}

\item{publish}{boolean; default: TRUE. Applies \code{\link[ggplot2]{theme_bw}} to the plot.}

\item{threshold}{numeric; default: 0. Minimal level of expression to be considered as an expressed gene}

\item{thm}{other theme object (\code{\link[ggplot2]{theme}}) to pass on to the plot}

\item{Y}{numeric vector of two elements. Specify the limites of the y-axis.
'c(minimal limit value, maximal limit value)'}

\item{remove.legend}{boolean; default: TRUE. Whether the legend should be removed
(do not impact the plot)}

\item{legendPosition}{either 'top','right','bottom' or 'left'. Where to position the legend.}

\item{out}{switch to chose the type of object returned by the function;
default 'DF', i.e. the data.frame with the formatted data used for the plotting.
Other options are:
                   'legend' for the legend of the plot,
                   'plot' for the figure itself,
                   'plot_noLegend' for the plot without the legend,
                   'plot_data' for the ggtable data to recreate the plot}

\item{print}{boolean; default: TRUE. Whether the plot should be printed}
}
\value{
depends on 'out' :
        'DF' returns the data.frame with all the required information for the plot,
        'legend' the legend of the plot,
        'plot' the figure itself,
        'plot_noLegend', the plot without the legend,
        'plot_data' the ggtable data to recreate the plot
}
\description{
Plot the distribution of genes across the columns.
Genes found uniquely in one condition are highlighted in another colour.
}
