% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{nl1d}
\alias{nl1d}
\title{Creates a two-column data.frame from a named vector where the column names are repeated in the first column}
\usage{
nl1d(filename)
}
\arguments{
\item{filename}{Name of the file which comprises a named vector}
}
\value{
a data.frame of two columns where the first column is the repetition of the row.names
}
\description{
Creates a two-column data.frame from a named vector where the column names are repeated in the first column
}
