% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{noMoreFactor}
\alias{noMoreFactor}
\title{Change all factor columns in a data.frame into character column}
\usage{
noMoreFactor(DF)
}
\arguments{
\item{DF}{a data.frame (with a factor column)}
}
\value{
a data.frame
}
\description{
Change all factor columns in a data.frame into character column
}
