% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{overlapSortCumul2DF}
\alias{overlapSortCumul2DF}
\title{Give the number of the common genes between two datasets at each rank}
\usage{
overlapSortCumul2DF(A, B, decreasing = TRUE)
}
\arguments{
\item{A}{Named numeric vector for the first dataset}

\item{B}{Named numeric vector for the second dataset}

\item{decreasing}{logical, default TRUE.
Whether the objects should be sorted prior to the comparison}
}
\value{
data.frame ready to be plotted with \code{\link[ggplot2]{ggplot}}-based function
}
\description{
Give the number of the common genes between two datasets at each rank
}
