% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{overlapSortCumul5DF}
\alias{overlapSortCumul5DF}
\title{Give the number of the common genes between five datasets at each rank}
\usage{
overlapSortCumul5DF(A, B, C, D, E, decreasing = TRUE)
}
\arguments{
\item{A}{Named numeric vector for the first dataset}

\item{B}{Named numeric vector for the second dataset}

\item{C}{Named numeric vector for the third dataset}

\item{D}{Named numeric vector for the fourth dataset}

\item{E}{Named numeric vector for the fifth dataset}

\item{decreasing}{logical, default TRUE.
Whether the objects should be sorted prior to the comparison}
}
\value{
data.frame ready to be plotted with \code{\link[ggplot2]{ggplot}}-based function
}
\description{
The genes need to be ranked based on
             a descriptor prior to the use of this function
}
