% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{overlapSpeProtmRNA}
\alias{overlapSpeProtmRNA}
\title{Overlap of the tissue-specific proteins with the most specific transcripts to that tissue.}
\usage{
overlapSpeProtmRNA(DFprot, DFmRNA, pretreatment = FALSE,
  mRNAmethod = "ratioSpe", cutoffProt = 0, complete.matrix = TRUE,
  pvalue = TRUE, ratio = 1, xprot = NA, yrna = NA, figure = TRUE,
  ...)
}
\arguments{
\item{DFprot}{numeric data.frame of the protein expression data}

\item{DFmRNA}{numeric data.frame of the transcriptomic expression data}

\item{pretreatment}{boolean. Default: FALSE. Whether the considered tissues (columns) and genes (rows)
need to be filtered to the common set.
The function requires the same set of tissues/genes to be meaningful.
This step can be bypassed if it already happened previously.}

\item{mRNAmethod}{a function name. The function should allow to order the mRNAs based on their specificity rank.}

\item{cutoffProt}{numeric. Threshold to consider a protein is expressed.}

\item{complete.matrix}{boolean. Default:TRUE.
Whether the comparison between the two studies should be done for their whole data.frame}

\item{pvalue}{boolean. Default: TRUE.
Whether the p-value of the jaccard index should be computed}

\item{ratio}{numeric (>0). Multiplying factor that
allows increasing the number of transcripts compared with the proteins.}

\item{xprot}{character string. Default: NA. Needed when complete.matrix is false.
Name of the tissue to consider for the proteomics.}

\item{yrna}{character string. Default: NA. Needed when complete.matrix is false.
Name of the tissue to consider for the transcriptomics.}

\item{figure}{boolean. Default: TRUE. Whether the venn diagram should be drawn}

\item{...}{more arguments for grDevices::cairo_pdf}
}
\value{
overlap between the proteomics and transcriptomics
}
\description{
Overlap of the tissue-specific proteins with the most specific transcripts to that tissue.
}
