% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptors.R
\name{plot_distribDF}
\alias{plot_distribDF}
\title{Plot the distribution for the columns of a data.frame according to different geometries (ggplot2)}
\usage{
plot_distribDF(DF, transformation = "log2", strip0 = TRUE,
  geom = "all", all = FALSE, debug = FALSE, verbose = FALSE,
  colourVal, limites, publish = TRUE, valueLab)
}
\arguments{
\item{DF}{numerical data.frame}

\item{transformation}{boolean; default: 'log2'. Whether if the data should be log2 transformed before being plotted}

\item{strip0}{boolean; default: TRUE; Whether the 0 should be stripped from the data before plotting}

\item{geom}{character string; default 'all'. Allows to chose between the different available geometries.
'all' output all available geometries. 'histogram' is for histogram, 'density' gives density plots.
Anything else gives a boxplot.}

\item{all}{boolean; default: TRUE. Whether all geometries should be plotted.}

\item{debug}{boolean. Default: FALSE}

\item{verbose}{boolean. Default: FALSE}

\item{colourVal}{character vector. Colour palette.}

\item{limites}{numeric vector. Allows to define the minimal and maximal boundaries to apply to the different plots.}

\item{publish}{boolean. Whether to apply  \code{\link[ggplot2]{theme_bw}} to the plot(s).}

\item{valueLab}{Character vector; allows to provide custom units if needed}
}
\value{
a distribution plot of the input data.frame
}
\description{
Plot the distribution for the columns of a data.frame according to different geometries (ggplot2)
}
