% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{plot_xsharedDistrib}
\alias{plot_xsharedDistrib}
\title{Plot the expression breadth distribution computed with xsharedDistrib}
\usage{
plot_xsharedDistrib(DF, colour, annotate = TRUE, publish = TRUE, ...)
}
\arguments{
\item{DF}{a numeric data.frame that comprises the expression breadth of the genes
in a column named 'nb.tissues'}

\item{colour}{colour palette for the different categories.
Default: "set2" of colorbrewer2}

\item{annotate}{boolean. Default: TRUE. Whether to annotate the figure
with the counts in each category for an accurate read.}

\item{publish}{boolean. Default: TRUE. Whether to apply ggplot2::theme_bw to the plot.}

\item{...}{other arguments that can be used by ggplot2::theme_bw()}
}
\value{
a figure
}
\description{
Plot the expression breadth distribution computed with xsharedDistrib
}
