% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-related.R
\name{plotGOHeatmap}
\alias{plotGOHeatmap}
\title{Create a heatmap from a compareClusterResult (\code{\link[clusterProfiler]{compareCluster}}) package)
showing for each enriched GO term, the overall observed correlation
of the associated genes set.}
\usage{
plotGOHeatmap(compareCluster, level = 5, corrValue, corr = TRUE,
  sortDF = TRUE, simpleHeatmap = TRUE,
  corName = "Pearson\\nCorrelation", base_family = "Linux Libertine",
  base_size = 16, stripName = FALSE, out = "DF", plot = TRUE)
}
\arguments{
\item{compareCluster}{Result object of \code{\link[clusterProfiler]{compareCluster}}}

\item{level}{integer; default: 5. How many level of the comparison between cluster should be retrieved}

\item{corrValue}{numeric vector. The correlation values for the different GO terms associated gene sets.}

\item{corr}{logical; default: TRUE. Whether the correlation should be plotted.}

\item{sortDF}{logical; default: TRUE. Whether the results should be sorted by correlation.}

\item{simpleHeatmap}{logical; default: TRUE.
Whether to remove the GO for which the gene set is empty.}

\item{corName}{string; default: "Pearson Correlation".
Allow to personalise the correlation legend}

\item{base_family}{string; default: "Linux Libertine". Name of the font family to use for the plot}

\item{base_size}{numeric; size of the main text of the plot; default: 16}

\item{stripName}{logical; default: FALSE.}

\item{out}{string; switch that allows to chose the type result object.
default:  "DF", gives a data.frame (used for the plot),
other options: "mat" gives the previous result formatted as a matrix
                "plot" gives the figure ready to be saved as an object.}

\item{plot}{logical; default: TRUE. Whether a plot should outputed to the standard output.}
}
\value{
depends of the value of "out"
}
\description{
Create a heatmap from a compareClusterResult (\code{\link[clusterProfiler]{compareCluster}}) package)
showing for each enriched GO term, the overall observed correlation
of the associated genes set.
}
