% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{plotTissuesDistance}
\alias{plotTissuesDistance}
\title{Plot heatmap based on the result of tissueDistance}
\usage{
plotTissuesDistance(DF, method = "ward.D", notecol = "gray37",
  col = colorRampPalette(c("aliceblue", "darkcyan")))
}
\arguments{
\item{DF}{numeric data.frame; output of tissueDistance}

\item{method}{character string, name of the agglomeration method to be used (hclust)
This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single",
"complete", "average" (= UPGMA), "mcquitty" (= WPGMA),
"median" (= WPGMC) or "centroid" (= UPGMC).}

\item{notecol}{character string. Name of the colour to use for the annotation within the heatmap.}

\item{col}{function to be used to define the colour palette for gplots::heatmap.2
default: colorRampPalette(c('aliceblue','darkcyan'))}
}
\value{
a heatmap (based on gplots::heatmap.2)
}
\description{
Plot heatmap based on the result of tissueDistance
}
