% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptors.R
\name{plotUniq.distrib}
\alias{plotUniq.distrib}
\title{More customisable plot based on computeBreadth}
\usage{
plotUniq.distrib(DF, omit.zero = TRUE, threshold = 0, annot = TRUE,
  publish = TRUE, ...)
}
\arguments{
\item{DF}{numeric data.frame}

\item{omit.zero}{numeric. Minimal expression to be considered.
For 0, the comparison is strict.}

\item{threshold}{numeric. Minimal expression to be considered.
For 0, the comparison is strict.}

\item{annot}{boolean. Default: TRUE.}

\item{publish}{boolean. Default: TRUE.
Whether to apply \code{\link[ggplot2]{theme_bw}} to the plot.}

\item{...}{allow to handle more theme elements in \code{\link[ggplot2]{theme_bw}}}
}
\value{
a plot
}
\description{
More customisable plot based on computeBreadth
}
