% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{prefix}
\alias{prefix}
\title{Prefix the names of the columns of a data.frame with a tag}
\usage{
prefix(DF, prefix, sep = ".")
}
\arguments{
\item{DF}{a data.frame}

\item{prefix}{a character vector; the tag to be added at the beginning
of each column of the data.frame}

\item{sep}{the separator between the tag and the original content; default='.'}
}
\value{
a data.frame with the exact same content
        and where the columns have been prefixed with a given tag
}
\description{
Prefix the names of the columns of a data.frame with a tag
}
