% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{prefixv}
\alias{prefixv}
\title{Prefix character vector with a given tag}
\usage{
prefixv(vec, prefix, sep = ".")
}
\arguments{
\item{vec}{a character vector}

\item{prefix}{the tag to be added at the beginning of each element of vec}

\item{sep}{the separator between the tag and the original content; default='.'}
}
\value{
a character vector
}
\description{
Prefix character vector with a given tag
}
