% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{presentData}
\alias{presentData}
\title{Wrapper that briefly present the datasets}
\usage{
presentData(DF, label.df, feature)
}
\arguments{
\item{DF}{a numeric (expression) dataset}

\item{label.df}{string of character that allows to label the dataset in the output}

\item{feature}{string of characters (essentially either 'gene', 'mRNA', 'RNA' or 'protein')}
}
\value{
a brief summary for the dataset
}
\description{
Wrapper that briefly present the datasets
}
