% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printDebug}
\alias{printDebug}
\title{For debugging purposes
Screen or report output
Look for a global variable DEBUG}
\usage{
printDebug(x)
}
\arguments{
\item{x}{the object to be printed}
}
\description{
For debugging purposes
Screen or report output
Look for a global variable DEBUG
}
