% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{protSpe}
\alias{protSpe}
\title{Give the list of rows for which the values are only once greater than the threshold}
\usage{
protSpe(DF, threshold = 0, type = "gt")
}
\arguments{
\item{DF}{a numeric data.frame}

\item{threshold}{numeric; default 0}

\item{type}{"gt" for greater than or anything else for greater than or equal to}
}
\value{
list of row names
}
\description{
Give the list of rows for which the values are only once greater than the threshold
}
