% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{rank.Atlas}
\alias{rank.Atlas}
\title{Rank for each tissue (column) the genes (rows)
based on their relative specificity to that tissue}
\usage{
rank.Atlas(DF, detect = 0)
}
\arguments{
\item{DF}{numeric data.frame}

\item{detect}{numeric. Threshold for detection; default: 0}
}
\value{
numeric data.frame
}
\description{
Rank for each tissue (column) the genes (rows)
based on their relative specificity to that tissue
}
