% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{rankSpe}
\alias{rankSpe}
\title{Rank gene expression based on their specificity to each tissue}
\usage{
rankSpe(DF, ties.method = "min", empty = "rank", order = TRUE,
  log2 = FALSE)
}
\arguments{
\item{DF}{numeric data.frame}

\item{ties.method}{how to handle ties. eg: 'min' or 'max'}

\item{empty}{how to handle missing data or bad score}

\item{order}{logical; default: TRUE. Whether the result should be order}

\item{log2}{logical; default:FALSE. Whether the data.frame should first be transformed}
}
\value{
numeric data.frame
}
\description{
Rank gene expression based on their specificity to each tissue
}
