% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{ratioSpe.1}
\alias{ratioSpe.1}
\title{Give the gene specificity in each tissue as a ratio to its mean across all tissue}
\usage{
ratioSpe.1(DF)
}
\arguments{
\item{DF}{numeric data.frame}
}
\value{
numeric data.frame
}
\description{
to avoid division by 0, 1 is added to all values.
}
