% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{ratioSpe}
\alias{ratioSpe}
\title{Compute the tissue specificity of the genes to each tissue as a ratio
to the sum of the gene expression across all tissues.}
\usage{
ratioSpe(DF)
}
\arguments{
\item{DF}{numeric data.frame}
}
\value{
numeric data.frame
}
\description{
Compute the tissue specificity of the genes to each tissue as a ratio
to the sum of the gene expression across all tissues.
}
