% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.list}
\alias{read.list}
\title{Read a file (one or two columns) and creates a vector (named if two-column file)}
\usage{
read.list(filename, header = TRUE, stringsAsFactors = FALSE,
  trunc = FALSE, ...)
}
\arguments{
\item{filename}{Name of the file}

\item{header}{boolean, TRUE if the file contains the header of the data}

\item{stringsAsFactors}{default FALSE}

\item{trunc}{default FALSE}

\item{...}{other parameters can be passed to utils::read.table}
}
\value{
a vector or a named vector if there are headers
}
\description{
Read a file (one or two columns) and creates a vector (named if two-column file)
}
