% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{Read.table}
\alias{Read.table}
\title{Basically a read.table but with predetermnined parameters}
\usage{
Read.table(filename, ..., path, row.names = 1)
}
\arguments{
\item{filename}{Name of the file of the data.frame to read}

\item{...}{any other parameter supported by utils::read.table}

\item{path}{Path of the file if not included in the name or not the working directory}

\item{row.names}{indicates the row that contains the column names}
}
\value{
an adequately formatted data.frame
}
\description{
Basically a read.table but with predetermnined parameters
}
