% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGplots.R
\name{redblue}
\alias{redblue}
\title{Generate red-to-blue colorscale}
\usage{
redblue(n)
}
\arguments{
\item{n}{Desired number of color elements in the panel.}
}
\value{
Vector of HTML-style RGB colors.
}
\description{
Generate red-to-blue colorscale
}
