% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGplots.R
\name{redgreen}
\alias{redgreen}
\title{Generate red-to-green colorscale}
\usage{
redgreen(n)
}
\arguments{
\item{n}{Desired number of color elements in the panel.}
}
\value{
Vector of HTML-style RGB colors.
}
\description{
Generate red-to-green colorscale
}
