% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{refGraphDF}
\alias{refGraphDF}
\title{Create a data.frame from the two data.frames after filtering the columns and rows.}
\usage{
refGraphDF(df1, df2, cond1, cond2 = cond1,
  select = intersect(row.names(df1), row.names(df2)))
}
\arguments{
\item{df1}{data.frame for a first study}

\item{df2}{data.frame for a second study}

\item{cond1}{vector of strings of character. Names of the columns to keep for df1}

\item{cond2}{vector of strings of character. Names of the columns to keep for df2.
default: "cond2=cond1"}

\item{select}{vector of strings of character. Names of the rows to keep in df1 and df2.
default: intersect(row.names(df1),row.names(df2)))}
}
\value{
a data.frame created from the junction of the df1 and df2
        after the filtering.
}
\description{
Create a data.frame from the two data.frames after filtering the columns and rows.
}
