% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{relative.GXenrichement2}
\alias{relative.GXenrichement2}
\title{Another function that annotate the genes that the expression is enriched of enhanced for a specific tissue}
\usage{
relative.GXenrichement2(DF, detect = 0, factors = FALSE)
}
\arguments{
\item{DF}{numeric data.frame}

\item{detect}{numeric; default:0. Minimal level of expression to be considered detected}

\item{factors}{logical. Whether the outputed classification should be characters or factors; default: FALSE}
}
\value{
data.frame
}
\description{
Another function that annotate the genes that the expression is enriched of enhanced for a specific tissue
}
