% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{relnum.GXenrichement}
\alias{relnum.GXenrichement}
\title{Another function that annotates the genes that the expression is enriched of enhanced for a specific tissue,
but the output is numeric}
\usage{
relnum.GXenrichement(DF, detect = 0)
}
\arguments{
\item{DF}{numeric data.frame}

\item{detect}{numeric; default:0. Minimal level of expression to be considered detected}
}
\value{
numeric data.frame
}
\description{
Another function that annotates the genes that the expression is enriched of enhanced for a specific tissue,
but the output is numeric
}
