% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{removeClusters}
\alias{removeClusters}
\title{Remove the proteins that are part of a cluster}
\usage{
removeClusters(DF, path)
}
\arguments{
\item{DF}{numeric (expression) data.frame}

\item{path}{path to the file that provide the clusters description}
}
\value{
a numeric data.frame where the problematic proteins have been removed
}
\description{
Remove the proteins that are part of a cluster
}
