% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{replotCumulSpeSimulAggregated}
\alias{replotCumulSpeSimulAggregated}
\title{Plot a figure based on the data.frame created by the cumulSpe-like function}
\usage{
replotCumulSpeSimulAggregated(DF_long, cor1, reference, Title,
  CorrelationPlot = TRUE, relabel = TRUE, cor2, simul1,
  aggregateBool = TRUE, aggregMethod = "rowMeans",
  labelVec = setNames(c("Protein/mRNA pairs",
  "mRNA/mRNA pairs ('ideal' reference)", "Randomised Protein/mRNA pairs"),
  c("Cor", "reference", "Simul")), sizeLine = 1, palette,
  legendText = 0.92, base_family = "Linux Libertine", base_size = 11,
  centerTitle = TRUE, out = "plot", VinterO8 = FALSE)
}
\arguments{
\item{DF_long}{data.frame in a long format}

\item{cor1}{named numeric vector that comprises the correlation between two studies.}

\item{reference}{named numeric vector that comprises the correlation
between two (transcriptomic) studies that can be used as a reference}

\item{Title}{character string. Title of the plot.}

\item{CorrelationPlot}{boolean. Default:TRUE. Whether to add the correlation plot to help for the interpretation.}

\item{relabel}{boolean. Default: TRUE. Whether the categories should be relaballed with the content of labelVec}

\item{cor2}{named numeric vector that comprises the correlation between two studies.}

\item{simul1}{named numeric vector that comprises the result of the random permutations on cor1}

\item{aggregateBool}{boolean. Default: TRUE. Whether to aggregate all the simulation of a given comparison together.}

\item{aggregMethod}{character string for the name of the method to aggregate the different simulation together.
Default: 'rowMeans'}

\item{labelVec}{named character string that allows to name the different categories to be displayed in the legend.
The names should be 'cor1, 'ref', 'simul1', .... and the content the labels
Default: "setNames(c("Protein/mRNA pairs",  "mRNA/mRNA pairs ('ideal' reference)",
          "Randomised Protein/mRNA pairs"), c('Cor','reference','simul1'))"}

\item{sizeLine}{positive numeric. Size of the lines of the plot}

\item{palette}{named vector of character strings. Allows to customise the colours on the plot.
Names should be identical to the categories or to the content of labelVec.}

\item{legendText}{positive numeric. Size of the legend text. Default: 0.92}

\item{base_family}{character string. Name of the font to use. Default: "Linux Libertine"}

\item{base_size}{positive numeric. Size of the font to use as a base.}

\item{centerTitle}{boolean. Default:TRUE. Whether to centre the title or keep it on the left.}

\item{out}{character string. Allows to chose the type of returned object.
"plot" (default) for the plot
and "annotatedPlot" for an annotated plot}

\item{VinterO8}{boolean. Default: FALSE. Add a vertical line for 80 percent of the data}
}
\value{
a figure
}
\description{
Plot a figure based on the data.frame created by the cumulSpe-like function
}
