% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{revCustom}
\alias{revCustom}
\title{Reverse a data.frame (or matrix) and change \code{-Inf} in \code{NA} or something else}
\usage{
revCustom(x, replacement = NA)
}
\arguments{
\item{x}{numeric data.frame or matrix}

\item{replacement}{numeric, default NA.}
}
\value{
numeric data.frame
}
\description{
Reverse a data.frame (or matrix) and change \code{-Inf} in \code{NA} or something else
}
