% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{rowCV}
\alias{rowCV}
\title{Covariance applied to a data.frame rows}
\usage{
rowCV(DF, digits = 1)
}
\arguments{
\item{DF}{a numeric data.frame}

\item{digits}{integer indicating the number of significant digits}
}
\value{
the covariance of each row (named vector)
}
\description{
cv = sd/mean
}
