% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{RowMad}
\alias{RowMad}
\title{Median absolute deviation applied to a data.frame rows}
\usage{
RowMad(DF, constant, ...)
}
\arguments{
\item{DF}{numeric data.frame}

\item{constant}{scale factor}

\item{...}{parameters that can be passed to stats::mad}
}
\value{
the m.a.d of each row (named vector)
}
\description{
Median absolute deviation applied to a data.frame rows
}
