% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{rowMax}
\alias{rowMax}
\title{Gives the maximum value of each row}
\usage{
rowMax(DF, type = "val")
}
\arguments{
\item{DF}{data.frame or matrix}

\item{type}{string to indicate how the result should be formatted
"val" (default) returns the maximum values,
"ind" returns the indices (as a named vector) of the maximum values,
"all" returns a data.frame with the maximum value for each row and the indices.}
}
\value{
the maximum values and (or) their indices
}
\description{
Gives the maximum value of each row
}
