% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{RowMeans}
\alias{RowMeans}
\title{Arithmetical mean applied to a data.frame rows}
\usage{
RowMeans(DF)
}
\arguments{
\item{DF}{numeric data.frame}
}
\value{
the mean of each row (named vector)
}
\description{
Arithmetical mean applied to a data.frame rows
}
