% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{RowMedians}
\alias{RowMedians}
\title{Arithmetical median applied to a data.frame rows}
\usage{
RowMedians(DF)
}
\arguments{
\item{DF}{numeric data.frame}
}
\value{
the median of each row (named vector)
}
\description{
Arithmetical median applied to a data.frame rows
}
