% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{rowMin}
\alias{rowMin}
\title{Gives the minimum value of each row}
\usage{
rowMin(DF, type = "val")
}
\arguments{
\item{DF}{data.frame or matrix}

\item{type}{string to indicate how the result should be formatted
"val" (default) returns the minimum values,
"ind" returns the indices (as a named vector) of the minimum values,
"all" returns a data.frame with the minimum value for each row and the indices.}
}
\value{
the minimum values and (or) their indices
}
\description{
Gives the minimum value of each row
}
