% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{RowMode}
\alias{RowMode}
\title{Mode applied to a data.frame rows}
\usage{
RowMode(DF, type = "mode")
}
\arguments{
\item{DF}{a numeric dataframe}

\item{type}{'mode' for the mode; anything else gives the frequence of the mode}
}
\value{
the mode (or its frequence) of each row (named vector)
}
\description{
Mode applied to a data.frame rows
}
