% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{rowSD}
\alias{rowSD}
\title{Standard deviation applied to a data.frame rows}
\usage{
rowSD(DF)
}
\arguments{
\item{DF}{a numeric data.frame}
}
\value{
standard deviation of each row (named vector)
}
\description{
Standard deviation applied to a data.frame rows
}
