% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{rowVar}
\alias{rowVar}
\title{Variance applied to a data.frame rows}
\usage{
rowVar(DF)
}
\arguments{
\item{DF}{a numeric data.frame}
}
\value{
variance of each row (named vector)
}
\description{
Variance applied to a data.frame rows
}
