% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{save_df}
\alias{save_df}
\title{Allow to save a data.frame quicly.}
\usage{
save_df(..., path = "../data/Robject", file = NA)
}
\arguments{
\item{...}{the data.frame to save}

\item{path}{character string. Path where to save the data.frame}

\item{file}{character string. Name of the file.
If left empty, a time stamped is used instead.}
}
\value{
the name of the saved file
}
\description{
Allow to save a data.frame quicly.
}
