% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{saveToFile}
\alias{saveToFile}
\title{Takes an object and saves it in a file with the same name.
Has an alternative where the name can be changed: saveWithName}
\usage{
saveToFile(obj, filename, path, overwrite, extension = ".RData",
  show.path = FALSE)
}
\arguments{
\item{obj}{the object to be saved}

\item{filename}{the filename can be specified. If it is not, the name object will be used}

\item{path}{the path where the object should be saved}

\item{overwrite}{if a file exists at the specified, if it should be overwrite with the new object}

\item{extension}{preferable to not change. By default: '.Rdata'}

\item{show.path}{Either if the path should be shown in the message. Default: `FALSE`}
}
\value{
message that indicates if the object has been saved (then with its name)
}
\description{
Takes an object and saves it in a file with the same name.
Has an alternative where the name can be changed: saveWithName
}
