% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{saveWithName}
\alias{saveWithName}
\title{Allows changing the name of the object at the same time that it is saved to a file}
\usage{
saveWithName(..., file)
}
\arguments{
\item{...}{object to be saved}

\item{file}{new name of the object which is saved with the same name}
}
\value{
saved file
}
\description{
Allows changing the name of the object at the same time that it is saved to a file
}
