% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter-plots.R
\name{scatplot}
\alias{scatplot}
\title{Scatter plot of two columns of a data.frame after possible log2 transformation.}
\usage{
scatplot(data, x, y, title, xlabs, ylabs, min, max, a = 0.65,
  log2 = TRUE, xcor = 3.3, ycor = 16.999, binx = 0.2, biny = 0.2,
  methodcor = "all", dens = TRUE, rug = TRUE, geomsize = 2,
  fig = "all", scaleCol, smooth = TRUE, xmin, ymin, xmax, ymax,
  abline = TRUE, publi = TRUE, col = NULL, labCol, verbose = FALSE)
}
\arguments{
\item{data}{numeric data.frame}

\item{x}{column name to be mapped to the x-axis}

\item{y}{column name to be mapped to the y-axis}

\item{title}{plot title}

\item{xlabs}{x-axis title}

\item{ylabs}{y-axis title}

\item{min}{numeric, minimum of both x-axis and y-axis}

\item{max}{numeric, maximum of both x-axis and y-axis}

\item{a}{numeric, alpha scaling factor to apply to the data}

\item{log2}{logical, default TRUE. Whether the data should be log2 transformed before plotting}

\item{xcor}{numeric, x coordinate for the correlation annotation to be anchored on its center}

\item{ycor}{numeric, y coordinate for the correlation annotation to be anchored}

\item{binx}{numeric, bin to use for the marginal plot of the x-axis}

\item{biny}{numeric, bin to use for the marginal plot of the y-axis}

\item{methodcor}{'all' for both Spearman and Pearson or or 'spearman' or 'pearson'}

\item{dens}{logical, default TRUE. Whether density countour curves should be added to the main figure}

\item{rug}{logical, default TRUE. Whether a rug should be added on the main figure}

\item{geomsize}{numeric, default 2. the size of the geom_point()}

\item{fig}{character string to chose from a set of predefined options that allows different outputs
'all' : complete figure with annotation and marginal plots
'grob': list of the grobs use by gridExtra::arrangeGrob to create the plot
'list': list of the different plots
'p'   : the main figure with the legend
'p1'  : the main figure without the legend
'p2'  : the marginal figure on the x-axis
'p3'  : the marginal figure on the y-axis
'legend': the legend of the figure only}

\item{scaleCol}{colour palette to use for the plot}

\item{smooth}{logical; default TRUE whether a smooth curve should be plotted}

\item{xmin}{numeric, minimum of the x-axis}

\item{ymin}{numeric, minimum of the y-axis}

\item{xmax}{numeric, maximum of the x-axis}

\item{ymax}{numeric, maximum of the y-axis}

\item{abline}{logical; default: TRUE; whether the line y=x should be added as a reference}

\item{publi}{logical, default: TRUE; whether to apply the theme_bw which is better suited for publication}

\item{col}{colour palette}

\item{labCol}{title for the colour scale}

\item{verbose}{logical; default FALSE}
}
\value{
a plot (see parameter fig for better description)
}
\description{
Scatter plot of two columns of a data.frame after possible log2 transformation.
}
