% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-genesExpDF.R
\name{selectSpecific}
\alias{selectSpecific}
\title{Select specifically the genes that are expressed above a threshold
for a given number of times in a data.frame}
\usage{
selectSpecific(DF, threshold = 0, expBreadth = 1, verbose = TRUE)
}
\arguments{
\item{DF}{numeric (expression) data.frame}

\item{threshold}{numeric; default: 0. Minimal expression to be considered}

\item{expBreadth}{integer; default: 1. Exact number of times to be expressed to be considered}

\item{verbose}{boolean; default: TRUE. Print the number of rows of the returned data.frame}
}
\value{
a data.frame that contains genes (ie rows) that
       are expressed exactly a number of times (expBreadth) above the given threshold
}
\description{
Select specifically the genes that are expressed above a threshold
for a given number of times in a data.frame
}
