% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setDebug}
\alias{setDebug}
\title{Allow to change the value of debug}
\usage{
setDebug(x)
}
\arguments{
\item{x}{boolean}
}
\value{
changes the value of sharedEnv$debug
}
\description{
Allow to change the value of debug
}
\examples{
setDebug(TRUE)
}
