% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setEpsilon}
\alias{setEpsilon}
\title{Allows to change the value of epsilon (needed to circumvent some R caveats)}
\usage{
setEpsilon(x)
}
\arguments{
\item{x}{New value for epsilon; default 1e-10}
}
\value{
Change the value of epsilon used by the other functions
}
\description{
Allows to change the value of epsilon (needed to circumvent some R caveats)
}
\examples{
setEpsilon(1e-9)
}
