% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{sharedBreadth}
\alias{sharedBreadth}
\title{Plot the breadth of a first study compared to a second one.}
\usage{
sharedBreadth(a, b, lapse, typeR = "vec", simplify, strict, Lims,
  colName = "nb.tissues")
}
\arguments{
\item{a}{numeric data.frame comprising the expression data of the first study.}

\item{b}{numeric data.frame comprising the expression data of the second study.}

\item{lapse}{positive integer.
Allow to relax constraints on perfect equality of breadth for the genes between the two studies.
A "similar" class is created. The lapse allows to define which level of similarity is acceptable.
e.g. if lapse=3, all genes that have a breadth of expression that varies at most of 3 are considered
to have similar breadth.}

\item{typeR}{character string. Allows to pick the type of output.
"vec" for a vector of the type of breath of expression observed in the first study compared to the second one.}

\item{simplify}{boolean. Whether to focus only on the most extreme breadth of expression.
"Expressed in all" or "nearly in all" might be more descriptive than "expressed in 5" versus "expressed in 6".}

\item{strict}{boolean. Whether if only strict equality between the breadth of expression of each gene between the two studies should be considered
or if the "lapse" argument should be used to define 'similar' cases as well.}

\item{Lims}{a length-4 numeric vector. When the analysis is in the "simplify" mode,
it allows to delimit two ranges of expression breadth to compare between the two studies.}

\item{colName}{character string. Gives the name of the column in the data.frame that has recorded the breath of expression of the genes}
}
\value{
depending on "typeR" either a vector or a data.frame
}
\description{
The colours refer to the breadth of expression in the other study.
}
