% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{signifDF}
\alias{signifDF}
\title{Round numbers to the specified number of significant digits to keep}
\usage{
signifDF(DF, i)
}
\arguments{
\item{DF}{numeric data.frame}

\item{i}{integer; indicates the number of significant digits to keep}
}
\value{
numeric data.frame
}
\description{
Round numbers to the specified number of significant digits to keep
}
