% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simpleCap2}
\alias{simpleCap2}
\title{Capitalise the first character of a string}
\usage{
simpleCap2(x)
}
\arguments{
\item{x}{a vector of characters}
}
\value{
the vector of characters with the first one capitalised
}
\description{
Capitalise the first character of a string
}
