% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{single.factoring}
\alias{single.factoring}
\title{From a file outputted by an older version of irap,
create a data.frame and add new columns with a selected generic value}
\usage{
single.factoring(filename, att, value, ...)
}
\arguments{
\item{filename}{character strings. Path of the file}

\item{att}{vector of attributes to add}

\item{value}{vector with the values with which to fill the new columns}

\item{...}{other arguments handled by multi.add.qual.att}
}
\value{
a data.frame
}
\description{
From a file outputted by an older version of irap,
create a data.frame and add new columns with a selected generic value
}
