% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{single.factoring2}
\alias{single.factoring2}
\title{From a filename, extract a data.frame to which a number of columns
(filled with selected values )}
\usage{
single.factoring2(filename, att, value, ...)
}
\arguments{
\item{filename}{character string, path to the file}

\item{att}{a vector of character strings (names of the new columns)}

\item{value}{vector of values for the new columns}

\item{...}{other parameters that can be handled by multi.add.qual.att}
}
\value{
a data.frame
}
\description{
From a filename, extract a data.frame to which a number of columns
(filled with selected values )
}
