% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-meta-analyses.R
\name{sortedGenesCorr}
\alias{sortedGenesCorr}
\title{Plot the correlation of different studies comparison that are ranked in decreasing order of correlation level.}
\usage{
sortedGenesCorr(cor1, cor2, ref, simul1, simul2, seed = 1235,
  aggregateBool = FALSE, aggregMethod = "rowMeans", title = "",
  xtitle = "Rank (sorted by decreasing order of correlation)",
  ytitle = "", unit = " (PPKM)", base_family = "Linux Libertine",
  base_size = 11, legendText = 0.92, ColPalette, geneList,
  out = "point")
}
\arguments{
\item{cor1}{named numeric vector. Comprises all the correlations between two studies for each genes.}

\item{cor2}{named numeric vector. Comprises all the correlations between two other studies for each genes.}

\item{ref}{named numeric vector. Comprises all the correlations between two studies that can be used as reference.}

\item{simul1}{named numeric vector or data.frame. Randomisation of cor1}

\item{simul2}{named numeric vector or data.frame. Randomisation of cor2}

\item{seed}{integer. Default: 1235. Allows to always pick the same simulation when data.frames are supplied.}

\item{aggregateBool}{boolean. Default:FALSE. Whether all the simulations should be aggreagated in one value for each gene per data.frame}

\item{aggregMethod}{character string for a funtion. Default:'rowMeans'.}

\item{title}{character string. Title of the plot.}

\item{xtitle}{character string. Label of the x-axis.}

\item{ytitle}{character string. Label of the y-axis.}

\item{unit}{character string. Unit in which the expression of the first study (Proteomics) is quantified}

\item{base_family}{character string. Name of the font to use.}

\item{base_size}{numeric. Size of the font to use (as a base)}

\item{legendText}{numeric. Size of the legend text.}

\item{ColPalette}{vector of character string to customise
the colours of "cor1", "cor2","simul1",...}

\item{geneList}{vector of character string. Gene identifiers that should be included in the figure.
default: common ids between cor1, cor2, simul1, simul2 and ref}

\item{out}{character string. Either "point" or "density"}
}
\value{
a figure
}
\description{
Plot the correlation of different studies comparison that are ranked in decreasing order of correlation level.
}
