% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{sortUhlenClassInOne}
\alias{sortUhlenClassInOne}
\title{Create one data.frame with the most relevant information for each gene
from previous gene classification ("Uhlén-like) analysis}
\usage{
sortUhlenClassInOne(notExpressed, threshold, NotDetected, Enriched,
  Enhanced, ubiHigh, ubiLow, mixedHigh, mixedLow, outfilename, gene.mapID)
}
\arguments{
\item{notExpressed}{object, or filename}

\item{threshold}{level of expression}

\item{NotDetected}{object, or filename}

\item{Enriched}{object, or filename}

\item{Enhanced}{object, or filename}

\item{ubiHigh}{object, or filename}

\item{ubiLow}{object, or filename}

\item{mixedHigh}{object, or filename for not expressed genes}

\item{mixedLow}{object, or filename for not expressed genes}

\item{outfilename}{filename for the resulting data.frame to be saved with}

\item{gene.mapID}{Map of the gene names with their Ensembl ids}
}
\value{
a data.frame with all the genes annotated with their "Uhlén-like" classification.
        result can be saved directly to a file simultaneously
}
\description{
Create one data.frame with the most relevant information for each gene
from previous gene classification ("Uhlén-like) analysis
}
