% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-deprecated.R
\name{stamped}
\alias{stamped}
\title{Attach the current time to a string name}
\usage{
stamped(string)
}
\arguments{
\item{string}{character string}
}
\value{
the string with the current time
}
\description{
Attach the current time to a string name
}
