% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stat.mode}
\alias{stat.mode}
\title{Statistical mode value}
\usage{
stat.mode(x, type = "mode")
}
\arguments{
\item{x}{a numeric vector}

\item{type}{mode' for the mode; anything else gives the frequence of the mode}
}
\value{
the mode (single numeric value) or its frequence
}
\description{
Compute the sample mode
             Nearly copy-paste from an answer from Ken Williams http://stackoverflow.com/users/169947/ken-williams
             see http://stackoverflow.com/a/8189441/2116422
}
