% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-filtering.R
\name{strip}
\alias{strip}
\title{Strip the rows of a data.frame that are failling the test.}
\usage{
strip(DF, test = "gt", threshold = 0, verbose = FALSE)
}
\arguments{
\item{DF}{data.frame to strip}

\item{test}{'gt' (greater than), 'ge' (greater than or equal),
'lt' (lesser than), 'le' (lesser than or equal)}

\item{threshold}{numeric value for the test}

\item{verbose}{when TRUE, shows the number of rows and columns before and after stripping}
}
\value{
a stripped data.frame
}
\description{
Strip the rows of a data.frame that are failling the test.
}
