% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{substrRight}
\alias{substrRight}
\title{Extract the last n character(s) from a string}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{a string}

\item{n}{number of characters to extract}
}
\value{
a vector of characters
}
\description{
from https://stackoverflow.com/a/7963963/2116422
}
