% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DF.R
\name{suffix}
\alias{suffix}
\title{Tag each column name of a data.frame with a given string of characters}
\usage{
suffix(DF, suf, sep = ".")
}
\arguments{
\item{DF}{a data.frame}

\item{suf}{the tag to be added at the end of each column name}

\item{sep}{the separator between the tag and the original content; default='.'}
}
\value{
a data.frame
}
\description{
Tag each column name of a data.frame with a given string of characters
}
