% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{t_test.CorrMat_identical.vs.different}
\alias{t_test.CorrMat_identical.vs.different}
\title{Test with a Welch t-test between the group of similar tissues versus the group of different tissues.}
\usage{
t_test.CorrMat_identical.vs.different(CorrMat, label.A, label.B,
  report = TRUE, index = "####", output = "DFtest_identical",
  publish = TRUE)
}
\arguments{
\item{CorrMat}{correlation matrix between two studies tissue expression (column correlation)}

\item{label.A}{character string. Name of the first study on which the correlations are based}

\item{label.B}{character string. Name of the second study on which the correlations are based}

\item{report}{boolean. Default: TRUE. Whether to output a text report.}

\item{index}{character string. For the notebook, allows to create new sections.}

\item{output}{character string.
'identical' to return only the pairs that comprise the same tissues,
'different' to return only the pairs that comprise the different tissues,
'DFtest' to return all the pairs,
'test' returns only the test,
'plot' returns only the plot,
'test_identical' returns the test and only the pairs that comprise the same tissues,
'DFtest_identical' returns all the pairs for the test and the identical tissues (as two different objects).}

\item{publish}{boolean. Default: TRUE. Whether to apply theme_bw()}
}
\value{
the result of the Welch t-test and a figure
}
\description{
Test with a Welch t-test between the group of similar tissues versus the group of different tissues.
}
