% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{t_test.Mat_identical.vs.different}
\alias{t_test.Mat_identical.vs.different}
\title{Test with a Welch t-test between the group of similar tissues versus the group of different tissues.
Take the diagonal and test it versus the lower triangle of the matrix.}
\usage{
t_test.Mat_identical.vs.different(Mat, yLab,
  label.diag = "Same-tissue pairs",
  label.low.tri = "Different tissues pairs", xLab = "Type",
  pval = TRUE, showMean = TRUE, output = "test", publish = TRUE)
}
\arguments{
\item{Mat}{correlation matrix between two studies tissue expression (column correlation)}

\item{yLab}{character string. Title of the y-axis}

\item{label.diag}{character string. Label for the pairs in the lower.triangle of the matrix.}

\item{xLab}{character string. Title of the x-axis. Default: 'Type'.}

\item{pval}{boolean. Default: TRUE. Whether to add the p-value on the plot.}

\item{showMean}{boolean. Default: TRUE. Whether to add the mean of each group on the plot.}

\item{output}{character string that allows to pick the output;
'identical' for the same pairs of tissues,
'different' for the different pairs of tissues,
'test' for the test,
'plot' for the plot and
'test_identical' for a list comprising the identical tissues and the test.}

\item{publish}{boolean. Default: TRUE. Whether to apply theme_bw().}
}
\value{
depends on 'output'; it can be a data.frame, a figure or the result of the test.
}
\description{
Test with a Welch t-test between the group of similar tissues versus the group of different tissues.
Take the diagonal and test it versus the lower triangle of the matrix.
}
