% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-related.R
\name{tableGeneIDdesc}
\alias{tableGeneIDdesc}
\title{Retrieve additional gene information from biomaRt}
\usage{
tableGeneIDdesc(geneID, ens, output = "DF", otherdata = NULL,
  NoIDcol = TRUE)
}
\arguments{
\item{geneID}{Ensembl gene ids to query}

\item{ens}{a "Mart" object from biomaRt}

\item{output}{"DF" for dataframe,"data.table" for a data.table::data.table,widget for a DT::datatable
and "latex" for a latex formatted output}

\item{otherdata}{other data that can be queried with biomaRt::getBM.
See possible attributes=biomaRt::listAttributes(ens)}

\item{NoIDcol}{logical; default=TRUE: remove the first column that contains the ensembl gene id}
}
\value{
a data.frame or a data.table or latex code or widget
}
\description{
Retrieve additional gene information from biomaRt
}
