% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{test_CorrMat_identical_different}
\alias{test_CorrMat_identical_different}
\title{Create the data.frame and can also test and plot
the comparison between the tissues of the first study with the second one}
\usage{
test_CorrMat_identical_different(CorrMat, test = FALSE, plot = FALSE,
  publish = TRUE, x = 0.75, y = 0.1, ym = 0.02, textaxis = 15,
  titleaxis = 15, out = "DFtest", sep = ".",
  remove.identity = TRUE, reorder = TRUE)
}
\arguments{
\item{CorrMat}{correlation matrix between two studies tissue expression (column correlation)}

\item{test}{boolean. Default: FALSE. Whether a (Welch's) t-test should be performed
between the group where the same tissue from each study are compared together
and the group where one tissue from the first study is compared to
a random tissue of the second tissue.}

\item{plot}{boolean. Default: FALSE. Whether the \strong{Test} should be plotted as well.}

\item{publish}{boolean. Default:TRUE. Whether the theme theme_bw() should be applied to the plot}

\item{x}{numeric. Default:0.75. x-axis of where to center the p-value.}

\item{y}{numeric. Default:0.1. y-axis of where to center the p-value.}

\item{ym}{numeric. Default: 0.02. y-axis of where to write the medium value of the correlation}

\item{textaxis}{numeric. Default:15. Size of the text on the x-axis}

\item{titleaxis}{numeric. Default:15. Size of the x-axis' title.}

\item{out}{character string. Allows to pick the output.
"DFtest" (default):}

\item{sep}{character or character string. Default: '.'}

\item{remove.identity}{boolean. Default:TRUE.
Whether the cases where the same tissue from the first study and the second one should be removed.}

\item{reorder}{boolean. Default: TRUE. Whether the data.frame should be order in decreasing order of the observed correlation}
}
\value{
depends on 'out'. Can be a data.frame, a plot, only the identical tissue comparisons,
        only the different tissue comparisons, ...
}
\description{
Create the data.frame and can also test and plot
the comparison between the tissues of the first study with the second one
}
