% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-extras.R
\name{theme_cbw}
\alias{theme_cbw}
\title{Theme black and white with centered title}
\usage{
theme_cbw(base_size = 11, base_family = "",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
Theme resetting the plot title at the center (as it was in the earlier version of ggplot2)
}
\examples{
ggplot(ggplot2::mpg, aes(cyl, hwy)) + geom_point() + ggtitle('Centered title') + theme_cbw()
}
\concept{themes cbw}
