% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-extras.R
\name{theme_minimaliste}
\alias{theme_minimaliste}
\title{Theme minimaliste: customised theme for ggplot2}
\usage{
theme_minimaliste(base_size = 11, base_family = "", ticks = TRUE)
}
\arguments{
\item{base_size}{Default: 11}

\item{base_family}{Default ""}

\item{ticks}{Boolean; whether the ticks should be drawn on the axes}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
Theme minimaliste: customised theme for ggplot2
}
\examples{
ggplot(ggplot2::mpg, aes(cyl, hwy)) + geom_point() + ggtitle('Centered title') + theme_minimaliste()
}
\concept{themes minimaliste}
