% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{tissuesDistance}
\alias{tissuesDistance}
\title{Compute the distance between the tissues within a study.}
\usage{
tissuesDistance(DF, nbOfTissues = 2, ColN, gid = FALSE, plot = FALSE,
  method = "ward.D", notecol = "gray37",
  col = colorRampPalette(c("aliceblue", "darkcyan")), threshold = 0,
  ...)
}
\arguments{
\item{DF}{numeric data.frame that contains}

\item{nbOfTissues}{integer (>0) number of tissues in which the genes have to be expressed to be considered}

\item{ColN}{vector of character strings. Names of the tissues to include in the analysis.}

\item{gid}{boolean. Default: FALSE. Whether to output the name of the concerned genes instead of their number.}

\item{plot}{boolean. Default: FALSE. Whether the heatmap should be printed or not}

\item{method}{character string, name of the agglomeration method to be used (hclust)
This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single",
"complete", "average" (= UPGMA), "mcquitty" (= WPGMA),
"median" (= WPGMC) or "centroid" (= UPGMC).}

\item{notecol}{character string. Name of the colour to use for the annotation within the heatmap.}

\item{col}{function to be used to define the colour palette for gplots::heatmap.2
default: colorRampPalette(c('aliceblue','darkcyan'))}

\item{threshold}{numeric. Default: 0}

\item{...}{for labelling purposes}
}
\value{
a data.frame
}
\description{
the distance is based on the number of genes that are uniquely shared by a number of given tissues only.
}
