% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{topCommonGenes}
\alias{topCommonGenes}
\title{Gives the number or the refenrences of shared genes
by both studies out of a given number of highest genes to query.}
\usage{
topCommonGenes(a, b, top = 0, filta, filtb, cutoff, LG = TRUE)
}
\arguments{
\item{a}{data.frame with expression data for the first study}

\item{b}{data.frame with expression data for the second study}

\item{top}{positive integer. Number of genes to consider for the analysis;
if 0 (default), all genes are considered.}

\item{filta}{character vector or positive integer vector.
rownames or indices of the rows to consider for the first study.}

\item{filtb}{character vector or positive integer vector.
rownames or indices of the rows to consider for the second study.}

\item{cutoff}{numeric. cutoff under which the expression is considered as noise}

\item{LG}{Boolean, topCommonGenes return the length if TRUE, else the list of genes is returned}
}
\value{
depends on LG. Number or the list of genes that are shared for the highest expressed
}
\description{
Gives the number or the refenrences of shared genes
by both studies out of a given number of highest genes to query.
}
