% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAnalyses.R
\name{topgenes.corCluster_log}
\alias{topgenes.corCluster_log}
\title{Compute the correlation matrix for the correlation of the two studies' tissues (columns)
after they have been transformed with log}
\usage{
topgenes.corCluster_log(a, b, filta = NA, same = FALSE, filtb = NA,
  top = 0, use = "pairwise.complete.obs", method = "pearson",
  nameA = deparse(substitute(a)), nameB = deparse(substitute(b)),
  cleanInfinite = TRUE, pseudocount = 1)
}
\arguments{
\item{a}{data.frame with expression data for the first study}

\item{b}{data.frame with expression data for the second study}

\item{filta}{character vector or positive integer vector.
colnames or indices of the columns to consider for the first study.}

\item{same}{boolean. Default:FALSE. Whether the tissues (columns)
from the second study should be identical to the first study.}

\item{filtb}{character vector or positive integer vector.
colnames or indices of the columns to consider for the second study.}

\item{top}{positive integer. Number of genes to consider for the analysis;
if 0 (default), all genes are considered.}

\item{use}{character string expliciting how missing values should be handled.
can be either  "everything", "all.obs", "complete.obs", "na.or.complete",
or "pairwise.complete.obs" (default).
To fasten the process if there is no missing value, use "everything".}

\item{method}{character string. Method for base::cor, either "pearson", "spearman" or
"kendall".}

\item{nameA}{character string to use for identifying the first study}

\item{nameB}{character string to use for identifying the second study}

\item{cleanInfinite}{boolean. Default: TRUE. whether to use clean.infinite to handle log(0)}

\item{pseudocount}{numeric. Default: 1. Pseudocount in case for log(x+pseudocount) when clean.infinite is not used for log(0)}
}
\value{
a matrix with the correlation of the columns of the two data.frames once log2-transformed
}
\description{
Compute the correlation matrix for the correlation of the two studies' tissues (columns)
after they have been transformed with log
}
