% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trimSpace}
\alias{trimSpace}
\title{Remove empty spaces in strings}
\usage{
trimSpace(x)
}
\arguments{
\item{x}{vector of characters}
}
\value{
the vector stripped of any space character
}
\description{
Remove empty spaces in strings
}
