% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{ubi.high}
\alias{ubi.high}
\title{Give the genes (rows) that are ubiquitously highly expressed across all tissues (columns)}
\usage{
ubi.high(DF, threshold = 10)
}
\arguments{
\item{DF}{numeric data.frame}

\item{threshold}{minimal threshold; default: 10}
}
\value{
vector with the names of the rows that check the condition
}
\description{
Give the genes (rows) that are ubiquitously highly expressed across all tissues (columns)
}
