% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{ubi.low}
\alias{ubi.low}
\title{Give the genes (rows) that are ubiquitously lowly expressed across all tissues (columns)}
\usage{
ubi.low(DF, threshold = 10, detect = 1)
}
\arguments{
\item{DF}{a numeric data.frame}

\item{threshold}{maximal threshold of expression; default: 10}

\item{detect}{minimal level of expression to be considered as expressed; default:1}
}
\value{
vector with the names of the rows that are
}
\description{
Give the genes (rows) that are ubiquitously lowly expressed across all tissues (columns)
}
