% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifications.R
\name{undetected}
\alias{undetected}
\title{List all the genes (rows) for which no expression was detected at a given threshold}
\usage{
undetected(DF, detect = 1)
}
\arguments{
\item{DF}{numeric data.frame}

\item{detect}{numeric; threshold of detection; default: 1}
}
\value{
vector of characters (gene names)
}
\description{
List all the genes (rows) for which no expression was detected at a given threshold
}
