% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Union}
\alias{Union}
\title{Allow to intersect more than two sets at a time}
\usage{
Union(...)
}
\arguments{
\item{...}{comma-separetated list of sets}
}
\value{
the union of all the sets
}
\description{
Allow to intersect more than two sets at a time
}
