% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{unmatchedRowsAbyB}
\alias{unmatchedRowsAbyB}
\title{Rows that are unmatched in one dataset by any row in another dataset.}
\usage{
unmatchedRowsAbyB(a, b)
}
\arguments{
\item{a}{first data.frame}

\item{b}{second data.frame}
}
\value{
Names of the unmatched rows
}
\description{
Rows that are unmatched in one dataset by any row in another dataset.
}
