% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{unmatchedRowsAbyBC}
\alias{unmatchedRowsAbyBC}
\title{Rows that are unmatched in one dataset by any row in two other datasets.}
\usage{
unmatchedRowsAbyBC(a, b, c)
}
\arguments{
\item{a}{first data.frame}

\item{b}{second data.frame}

\item{c}{third data.frame}
}
\value{
Names of the unmatched rows
}
\description{
Rows that are unmatched in one dataset by any row in two other datasets.
}
