% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{usablePath}
\alias{usablePath}
\title{Check if path is available for write}
\usage{
usablePath(p)
}
\arguments{
\item{p}{path that has to be check for writing permission}
}
\value{
adequate path
}
\description{
Check if path is available for write
}
