% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiag.R
\name{vennPlot_spe}
\alias{vennPlot_spe}
\title{Compare the content of the two first columns of a data.frame
after the application of a threshold}
\usage{
vennPlot_spe(DF, cutoff = 0.5, fill = c("red", "blue"))
}
\arguments{
\item{DF}{a numeric data.frame}

\item{cutoff}{a threshold above which the values should be observed}

\item{fill}{a vector with two colours}
}
\value{
a venn diagram
}
\description{
Compare the content of the two first columns of a data.frame
after the application of a threshold
}
