% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-helpers.R
\name{wrap.scatplot.log2}
\alias{wrap.scatplot.log2}
\title{Wrapper automating the creation of scatter plots (scatplot.log2)
based on scatplot.log2}
\usage{
wrap.scatplot.log2(DFindex, DF1, DF2, label.df1, label.df2, out.df1,
  out.df2, xmin, xmax, ymin, ymax, a, xcor, ycor, methodcor, fig,
  abline = FALSE, verbose = TRUE, publi = TRUE, report = TRUE,
  pdf = FALSE)
}
\arguments{
\item{DFindex}{When used for the output for "identical" (t.test output) should be already sorted}

\item{DF1}{first numeric data.frame}

\item{DF2}{second numeric data.frame}

\item{label.df1}{name on the title for DF1 (and used to create the filename)}

\item{label.df2}{name on the title for DF2 (and used to create the filename)}

\item{out.df1}{name to output on the x-axis (for DF1)}

\item{out.df2}{name to output on the y-axis (for DF2)}

\item{xmin}{lower limit for the x-axis}

\item{xmax}{upper limit for the x-axis}

\item{ymin}{lower limit for the y-axis}

\item{ymax}{upper limit for the y-axis}

\item{a}{numeric; alpha scale}

\item{xcor}{numeric; x coordinate to center the correlation}

\item{ycor}{numeric; y coordinate to center the correlation}

\item{methodcor}{"spearman" or "pearson"}

\item{fig}{switch; "grob"; "all"}

\item{abline}{logical; default: FALSE. Whether x=y should be drawn.}

\item{verbose}{logical; default: TRUE.}

\item{publi}{logical; default: TRUE}

\item{report}{logical; default: TRUE}

\item{pdf}{logical; default: FALSE}
}
\value{
several scatter plots outputted as a pdf or part as html report
}
\description{
Wrapper automating the creation of scatter plots (scatplot.log2)
based on scatplot.log2
}
