% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{wtDF}
\alias{wtDF}
\title{write.table applied to DF plus some general customisation}
\usage{
wtDF(DF, filename, path, rowNames = "Gene.ID.ens76", overwrite,
  extension = ".tsv", show.path = FALSE)
}
\arguments{
\item{DF}{data.frame to be saved to a text file}

\item{filename}{Name of the file to be saved}

\item{path}{Path where the file should be saved}

\item{rowNames}{The name of the first column in the file}

\item{overwrite}{if `TRUE` indicates that the file should be saved in any case,
even if it overwrites a previous file with the same name at the same path}

\item{extension}{by default '.tsv'}

\item{show.path}{indicates if the path in the message should be printed too}
}
\value{
message that indicates if the object has been saved to a (named) text file
}
\description{
write.table applied to DF plus some general customisation
}
